% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_tplotfunctions.R
\name{alpha2hex2}
\alias{alpha2hex2}
\title{Various plotting and statistics functions}
\usage{
alpha2hex2(alpha, col = NULL)
}
\description{
Various plotting and statistics functions
}
\keyword{internal}
