% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmetadata.R
\name{buildmetadata}
\alias{buildmetadata}
\alias{metadata}
\title{Metadata}
\usage{
buildmetadata(
  data,
  file = NULL,
  includevrt = NULL,
  excludevrt = NULL,
  addsummary2metadata = FALSE,
  backupfiles = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{A dataset, given as a \code{\link[base]{data.frame}}
or as a file path to a csv file.}

\item{file}{String: name of csv file where the metadata should be saved;
if \code{NULL}: output metadata as \code{VALUE}.}

\item{includevrt}{Character or \code{NULL}: name of variates in dataset to be included.}

\item{excludevrt}{Character or \code{NULL}: name of variates in dataset to be excluded.}

\item{addsummary2metadata}{Logical: also output some diagnostic statistics
in the metadata? Default \code{FALSE}.}

\item{backupfiles}{Logical: rename previous metadata file if it exists?
Default \code{TRUE}.}

\item{verbose}{Logical: output heuristics for each variate? Default \code{TRUE}.}
}
\value{
If \code{file = NULL}, a preliminary metadata file is created
and \code{VALUE} is \code{NULL};
otherwise \code{VALUE} is a \code{\link[base]{data.frame}} containing the metadata.
}
\description{
Details about metadata, metadata file and object, and utility function to build a metadata file or object.
}
\details{
This is a first test
}
\section{Necessity of metadata}{


This is a test
}

