% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdaptiveMCMCSampler.R
\name{Criterion}
\alias{Criterion}
\title{R6 class defining a criterion for a sampler}
\description{
A selection criterion object can determine if a given proposal is
   accepted given the probability of the proposal and a reference probability

   This class is abstract and is not intended to be instantiated
   directly. Inheriting parameter processors are specific to a given 
   type of model input and the model that takes that input before execution.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-isAccepted}{\code{Criterion$isAccepted()}}
\item \href{#method-clone}{\code{Criterion$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isAccepted"></a>}}
\if{latex}{\out{\hypertarget{method-isAccepted}{}}}
\subsection{Method \code{isAccepted()}}{
Determine if the provided probability is acceptable or not

  This method is abstract, indicating that implementations of 
  this class need to override this method or an error will be
  thrown when it is called.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Criterion$isAccepted(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{}

\item{\code{prob}}{Probability being evaluated}

\item{\code{probRef}}{Reference probability for comparison}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TRUE if accepted, FALSE otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Criterion$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
