% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E.R
\name{E}
\alias{E}
\title{Calculate posterior probabilities}
\usage{
E(
  Ynames,
  X = NULL,
  learnt,
  quantiles = c(5, 95)/100,
  nsamples = 100L,
  parallel = TRUE,
  silent = FALSE,
  usememory = TRUE
)
}
\arguments{
\item{X}{matrix or data.table or \code{NULL}: set of values of variates on which we want to condition the joint probability of \code{Y}. If \code{NULL} (default), no conditioning is made (except for conditioning on the learning dataset and prior assumptions). One variate per column, one set of values per row.}

\item{learnt}{Either a string with the name of a directory or full
path for an 'learnt.rds' object, or such an object itself}

\item{quantiles}{numeric vector, between 0 and 1, or \code{NULL}: desired quantiles of the variability of the probability for \code{Y}. Default \code{c(0.05, 0.95)} or the 5\% and 95\% quantiles.}

\item{nsamples}{integer or \code{NULL}: desired number of samples of the variability of the probability for \code{Y}. Default \code{100}.}

\item{parallel}{logical or integer: whether to use pre-existing parallel
workers, or how many to create and use. Default \code{TRUE}.}

\item{silent}{logical: give warnings or updates in the computation?
Default \code{FALSE}.}

\item{usememory}{logical: save partial results to disc, to avoid crashes?
Default \code{TRUE}.}

\item{Y1names}{String vector: joint variates
the joint probability of. One variate per column, one set of values per row.}
}
\value{
A list of: (1) a matrix with the probabilities P(Y|X,data,assumptions), for all combinations of values of \code{Y} (rows) and \code{X} (columns); (2) an array with the variability quantiles (3rd dimension of the array) for such probabilities; (3) an array with the variability samples (3rd dimension of the array) for such probabilities.
}
\description{
This function calculates the probability P(Y | X, data), where Y and X are two (non overlapping) sets of joint variates. The function also gives quantiles about the possible variability of the probability P(Y | X, newdata, data) that we could have if more learning data were provided, as well as a number of samples of the possible values of such probabilities. If several joint values are given for Y or X, the function will create a 2D grid of results for all possible compbinations of the given Y and X values.
}
