% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_mcmcstop.R
\name{fftNGS}
\alias{fftNGS}
\title{Find optimal FFT size}
\usage{
fftNGS(N)
}
\arguments{
\item{N:}{integer.}
}
\value{
Optimal FFT size
}
\description{
From \strong{rstan} \url{https://github.com/stan-dev/rstan/blob/develop/rstan/rstan/R/monitor.R}
}
\keyword{internal}
