% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_mcmcstop.R
\name{funESS}
\alias{funESS}
\title{Function for calculating MC effective sample size.}
\usage{
funESS(x)
}
\arguments{
\item{x:}{a matrix, rows being MC samples and columns being quantities whose MCSE is to be estimated.}
}
\value{
Estimates of the effective sample size for each trace.
}
\description{
Modified from
from https://github.com/LaplacesDemonR/LaplacesDemon/blob/master/R/ESS.R
}
\keyword{internal}
