% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_mcmcstop.R
\name{funESS2}
\alias{funESS2}
\title{Function for calculating MC standard error, from Geyer's mcmc package}
\usage{
funESS2(x)
}
\arguments{
\item{x:}{a matrix, rows being MC samples and columns being quantities whose ESS is to be estimated.}
}
\value{
Estimates of ESS for each trace.
}
\description{
Function for calculating MC standard error, from Geyer's mcmc package
}
\keyword{internal}
