% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_mcmcstop.R
\name{funMCSELD}
\alias{funMCSELD}
\title{Calculate MC standard error using LaplacesDemon's batch means}
\usage{
funMCSELD(x)
}
\arguments{
\item{x}{matrix, each row being a "trace", that is a set of MC samples, whose MCSE is to be estimated.}
}
\value{
MCSE estimates, one for each trace. Division by sqrt(N) is already performed.
}
\description{
Modified from
from https://github.com/LaplacesDemonR/LaplacesDemon/blob/master/R/ESS.R
}
\details{
Used in 'util_combineYX()' in 'Pr()'.
}
\keyword{internal}
