% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_datacheck.R
\name{inferno.data}
\alias{inferno.data}
\alias{write.csvi}
\alias{read.csvi}
\title{Write and read data in \strong{inferno}}
\usage{
write.csvi(x, file, ...)

read.csvi(file, ...)
}
\arguments{
\item{x}{the object to be written, preferably a matrix or data frame. If not, it is attempted to coerce \code{x} to a data frame. See \code{\link[utils]{write.table}}.}

\item{file}{either a character string naming a file or a connection open for writing or reading. See \code{\link[utils]{write.table}} and \code{\link[utils]{read.table}}.}

\item{...}{Other parameters to be passed to \code{\link[utils]{write.table}} or \code{\link[utils]{read.table}}.}
}
\description{
Utility functions to read and write CSV files in the format required by \strong{inferno}
}
\details{
The functions of the \strong{inferno} package accept CSV files formatted as follows:
\itemize{
\item Decimal values should be separated by a \emph{dot}; no comma should be used to separate thousands etc. Example: \code{86342.75} .
\item Character and names should be quoted in single or double quotes. Example: \code{"female"}.
\item Values should be separated by \emph{commas}, not by tabs or semicolons.
\item Missing values should be simply \emph{empty}, not denoted by "NA", "missing", "-", or similar.
\item Preferably there should not be \code{\link[base]{factor}}s; use character names instead.
}

The utility functions \code{\link[=write.csvi]{write.csvi()}} and \code{\link[=read.csvi]{read.csvi()}} are wrappers to \code{\link[utils]{write.table}} or \code{\link[utils]{read.table}} that sets appropriate default parameters
}
