% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferpopulation.R
\name{inferpopulation}
\alias{inferpopulation}
\title{Monte Carlo computation of posterior distribution of population frequencies}
\usage{
inferpopulation(
  data,
  metadata,
  auxdata = NULL,
  outputdir,
  nsamples = 1200,
  nchains = 120,
  nsamplesperchain = 10,
  parallel = TRUE,
  seed = NULL,
  cleanup = TRUE,
  appendtimestamp = TRUE,
  appendinfo = TRUE,
  subsampledata = NULL,
  output = "directory",
  niterini = 1200,
  miniter = 3600,
  maxiter = +Inf,
  prior = missing(data),
  thinning = NULL,
  plottraces = TRUE,
  showKtraces = FALSE,
  showAlphatraces = FALSE,
  lldata = 12
)
}
\arguments{
\item{data}{data.frame object or filepath: datapoints}

\item{metadata}{Either the name of the csv file containing metadata
of the current dataset, or a data.frame with the metadata}

\item{auxdata}{data.frame object or filepath: extra datapoints
which would be too many to use in the Monte Carlo sampling,
but can be used to calculate hyperparameters}

\item{outputdir}{String, path to output file folder ## Rename to
outputPrefix, also addSuffix?}

\item{nsamples}{Integer, nr of desired MC samples}

\item{nchains}{Integer, nr of MC chains}

\item{nsamplesperchain}{Integer, nr of MC samples per chain}

\item{parallel, }{logical or numeric: whether to use pre-existing parallel
workers, or how many to create and use}

\item{seed}{Integer: seed for random number generator. If left as default
NULL, a random seed based on the system clock is used in the
set.seed() function}

\item{cleanup}{logical, default TRUE, removes files that can be used for
debugging}

\item{subsampledata}{Numeric: use only a subsample of the datapoints in 'data'}

\item{output}{if string 'directory', return the output directory name;
if string 'mcoutput', return the 'Fdistribution' object;
anything else, no output}

\item{niterini}{Number of initial (burn-in) MC iterations}

\item{miniter}{Minimum number of MC iterations to be done}

\item{maxiter}{Maximum number of MC iterations}

\item{prior}{logical: Calculate the prior distribution of F?}

\item{thinning}{If NULL, let the diagnostics decide the MC thinning; if positive, use this thinning value}

\item{plottraces}{logical: plot MC traces of diagnostic values}

\item{showKtraces}{logical, when true, it saves the K parameter during
sampling and plots its trace and histogram at the end. Keeping it to
FALSE (default) saves a little computation time.}

\item{showAlphatraces}{logical, : when true, it saves the Alpha parameter
more frequently during sampling and plots its trace and histogram at
the end. Keeping it to FALSE (default) saves a little computation
time.}

\item{lldata}{Positive integer or Inf (default 12): number of datapoints
to use for loglikelihood calculations; if Inf, use all.}
}
\value{
name of directory containing output files, or Fdistribution object, or empty
}
\description{
Monte Carlo computation of posterior distribution of population frequencies
}
