% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_mcsamples.R
\name{learnbind}
\alias{learnbind}
\title{Bind 3D arrays by first dimension}
\usage{
learnbind(x, y)
}
\description{
Used in 'util_checkpoints()' within 'learn()', and in various functions in 'util_lprobs.R'.
}
\details{
NB: the following variant is slower:

\if{html}{\out{<div class="sourceCode">}}\preformatted{function(x, y) \{
    out <- c(aperm(x), aperm(y))
    dim(out) <- c(rev(dim(x)[-1]), dim(x)[1] + dim(y)[1])
    aperm(out)
\}
}\if{html}{\out{</div>}}
}
\keyword{internal}
