% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_mcmcstop.R
\name{mcmcstop}
\alias{mcmcstop}
\title{Function for calculating the number of needed MCMC iterations}
\usage{
mcmcstop(traces, nsamples, availiter, relerror, thinning)
}
\description{
Function for calculating the number of needed MCMC iterations
}
