% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadataExample}
\alias{metadataExample}
\title{Example metadata file}
\format{
\subsection{\code{metadataExample}}{

A data frame with 2 rows and 10 columns.
}
}
\usage{
metadataExample
}
\description{
A data frame containing the prior information about the variates \code{species} and \code{bill_len} of the \link[datasets:penguins]{datasets::penguins} dataset.
}
\seealso{
\code{\link[=metadatatemplate]{metadatatemplate()}} which helps producing this metadata file from a given dataset, \code{\link[=learn]{learn()}} which uses to produce hyperparameters for posterior infereces, \code{learntExample} as an (unrealistic) `learnt' object produced from these metadata.
}
\keyword{datasets}
