% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{Model}
\alias{Model}
\title{R6 Abstract class defining a model interface}
\description{
Inferno classes use a Model interface to manipulate and run models.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Model-run}{\code{Model$run()}}
\item \href{#method-Model-clone}{\code{Model$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-run"></a>}}
\if{latex}{\out{\hypertarget{method-Model-run}{}}}
\subsection{Method \code{run()}}{
A class that inherits from Model must implement a "run" method that will
  execute the model with its given configuration.

  This method declaration will cause a program to fail if 
  the method is called and the implementing class does not override it.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$run()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
No required return value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Model-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Model-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Model$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
