% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObservationGenerator.R
\name{ObservationGeneratorNormalErr}
\alias{ObservationGeneratorNormalErr}
\title{R6 class defining an observation generator assuming normal error}
\description{
A observation generator for independent and normally
  distributed error.
}
\section{Super class}{
\code{\link[inferno:ObservationGenerator]{inferno::ObservationGenerator}} -> \code{ObservationGeneratorNormalErr}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mean}}{Mean of the normal distribution of artificial error}

\item{\code{sd}}{Standard deviation of the distribution of artificial error}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ObservationGeneratorNormalErr-new}{\code{ObservationGeneratorNormalErr$new()}}
\item \href{#method-ObservationGeneratorNormalErr-generate}{\code{ObservationGeneratorNormalErr$generate()}}
\item \href{#method-ObservationGeneratorNormalErr-clone}{\code{ObservationGeneratorNormalErr$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObservationGeneratorNormalErr-new"></a>}}
\if{latex}{\out{\hypertarget{method-ObservationGeneratorNormalErr-new}{}}}
\subsection{Method \code{new()}}{
Construct a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationGeneratorNormalErr$new(mean, sd)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mean}}{Mean of the normal distribution to use for artificial error}

\item{\code{sd}}{Standard deviation of the normal distribution to use for
artifical error}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObservationGeneratorNormalErr-generate"></a>}}
\if{latex}{\out{\hypertarget{method-ObservationGeneratorNormalErr-generate}{}}}
\subsection{Method \code{generate()}}{
Generates synthetic observations based on normally-distributed
  and independent error.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationGeneratorNormalErr$generate(base)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base}}{Basis for the synthetic observations. Normally distributed error
is added to the basis to create the observations.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object representing the synthetic observations
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObservationGeneratorNormalErr-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ObservationGeneratorNormalErr-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationGeneratorNormalErr$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
