% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tplotfunctions.R
\name{pdff}
\alias{pdff}
\title{Various plotting and statistics functions}
\usage{
pdff(
  file = "Rplot",
  apaper = 5,
  portrait = FALSE,
  height = 148/25.4,
  width = 210/25.4,
  asp = NA,
  ...
)
}
\description{
Various plotting and statistics functions
}
