% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_tplotfunctions.R
\name{plot.probability}
\alias{plot.probability}
\title{Plot an object of class "probability"}
\usage{
\method{plot}{probability}(
  p,
  variability = NULL,
  PvsY = NULL,
  legend = TRUE,
  lty = c(1, 2, 4, 3, 6, 5),
  lwd = 2,
  col = c("#4477AA", "#EE6677", "#228833", "#CCBB44", "#66CCEE", "#AA3377"),
  xlab = NULL,
  ylab = NULL,
  ylim = c(0, NA),
  grid = TRUE,
  add = FALSE,
  ...
)
}
\arguments{
\item{p}{object of class "probability", obtained with \code{\link{Pr}} or \code{\link{tailPr}}.}

\item{variability}{one of the values \code{"quantiles"}, \code{"samples"}, \code{"none"} (equivalent to \code{NA} or \code{FALSE}), or \code{NULL} (default), in which case the variability available in \code{p} is used. This argument chooses how to represent the variability of the probability; see \code{\link{Pr}}. If the requested variability is not available in the object \code{p}, then a warning is issued and no variability is plotted.}

\item{PvsY}{logical or \code{NULL}: should probabilities be plotted against their \code{Y} argument? If \code{NULL}, the argument between \code{Y} and \code{X} having larger number of values is chosen. As many probability curves will be plotted as the number of values of the other argument.}

\item{legend}{string or logical: plot a legend of the different curves at position \code{legend}? If \code{TRUE}, position is 'top'.}

\item{...}{other parameters to be passed to \code{\link[base]{matplot}}.}
}
\description{
This plot method is a utility to plot probabilities obtained with \code{\link{Pr}} or \code{\link{tailPr}}, as well as their uncertainties. The probabilities are plotted either against \code{Y}, with one curve for each value of \code{X}, or vice versa.
}
