% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFsamples.R
\name{plotFsamples}
\alias{plotFsamples}
\title{Plot one-dimensional posterior probabilities}
\usage{
plotFsamples(
  file,
  mcoutput,
  data,
  plotmeans = TRUE,
  plotvariability = "samples",
  nsamples = 100,
  datahistogram = TRUE,
  datascatter = TRUE,
  useOquantiles = FALSE,
  parallel = TRUE,
  silent = FALSE
)
}
\arguments{
\item{file}{string: name of plot output file}

\item{mcoutput}{Either a string with the name of a directory or full
path for a 'FDistribution.rds' object, or such an object itself}

\item{data}{data.table object or filepath: datapoints}

\item{plotvariability}{string, either 'samples' or 'quantiles': how to plot the variability of the probability distribution with new samples}

\item{nsamples}{positive number: number of samples of representative frequency distributions to display as variability}

\item{datahistogram}{bool: plot the data as histogram?}

\item{datascatter}{bool: plot the data as scatterplot along the x-axis?}

\item{useOquantiles}{Bool: internal, use metadata quantiles for ordinal
variates}

\item{parallel}{Bool or numeric: whether to use pre-existing parallel
workers, or how many to create and use}

\item{silent}{Bool: give warnings or updates in the computation}
}
\value{
A list with the mutual information, its error, and its unit
}
\description{
Plot one-dimensional posterior probabilities
}
