% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictionExtractor.R
\name{PredictionExtractor}
\alias{PredictionExtractor}
\title{R6 class defining a simulation prediction extractor}
\description{
Extracts a relatively simple subset of model ouptut used to compare 
  the model prediction to other data. For example, a prediciton extractor may
  be used to extract a simple vector of predicted values to compare
  to observed values in an inferential modeling algorithm.

  This class is abstract and is not intended to be instantiated
  directly. Inheriting parameter extractors are specific to a given 
  type of model input and the model that takes that input before execution.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model}}{Model object from which the output is extracted}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PredictionExtractor$new()}}
\item \href{#method-setModel}{\code{PredictionExtractor$setModel()}}
\item \href{#method-extract}{\code{PredictionExtractor$extract()}}
\item \href{#method-clone}{\code{PredictionExtractor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Constructs a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExtractor$new(model = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{The model from which the output is extracted}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setModel"></a>}}
\if{latex}{\out{\hypertarget{method-setModel}{}}}
\subsection{Method \code{setModel()}}{
Set the model from which predictions are extracted
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExtractor$setModel(model)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{The model with output that can be extracted as a
prediction by the extract method.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extract"></a>}}
\if{latex}{\out{\hypertarget{method-extract}{}}}
\subsection{Method \code{extract()}}{
A class that inherits from PredictionExtractor must implement an 
  "extract" method that will extract the desired subset of model
  output. 

  This method declaration will cause a program to fail if 
  the method is called and the implementing class does not override it.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExtractor$extract()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An object representing the subset of model output
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PredictionExtractor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
