% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferno-package.R
\docType{package}
\name{prova-package}
\alias{prova}
\alias{prova-package}
\title{prova: Probabilistic-Statistical Variate Analysis and Inference with Bayesian Nonparametrics}
\description{
Functions for calculating posterior joint and conditional probabilities, probability distributions of population frequencies, and information-theoretic measures, by means of Bayesian nonparametric methods. Markov-chain Monte Carlo calculations are automatically handled and do not require user supervision. Applications range from statistical estimation and probabilistic hypothesis testing to evidence-based inference and decision making, in a wide range of disciplines from astrophysics to medicine. For more details and examples see for instance Porta Mana et al. (2026) \doi{10.31219/osf.io/8nr56}, Dunson & Bhattacharya (2011) \doi{10.1093/acprof:oso/9780199694587.003.0005}, Müller et al. (2015) \doi{10.1007/978-3-319-18968-0}, Hjort et al. (2010) \doi{10.1017/CBO9780511802478}. Requires the packages 'Nimble', 'parallel', 'extraDistr'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pglpm.github.io/prova/}
  \item \url{https://github.com/pglpm/prova/}
}

}
\author{
\strong{Maintainer}: PierGianLuca Porta Mana \email{pgl@portamana.org} (\href{https://orcid.org/0000-0002-6070-0784}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item PierGianLuca Porta Mana \email{pgl@portamana.org} (\href{https://orcid.org/0000-0002-6070-0784}{ORCID}) [copyright holder]
  \item Aurora Grefsrud \email{agre@hvl.no} (\href{https://orcid.org/0009-0005-9063-4131}{ORCID})
  \item Håkon Mydland \email{haakon.mydland@gmail.com} (\href{https://orcid.org/0009-0004-9558-8894}{ORCID})
  \item Maksim Ohvrill \email{maksimohvrill@hotmail.com}
  \item Simen Hesthamar Hauge \email{simen@hhnet.no} (\href{https://orcid.org/0009-0005-6158-2846}{ORCID})
}

}
\keyword{internal}
