% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomVariable.R
\name{RandomVariable}
\alias{RandomVariable}
\title{R6 class defining a random variable}
\description{
A class representing a random variable with a given distribution.
   This class is abstract and is not intended to be instantiated
   directly.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RandomVariable-density}{\code{RandomVariable$density()}}
\item \href{#method-RandomVariable-markovStep}{\code{RandomVariable$markovStep()}}
\item \href{#method-RandomVariable-normalizedFit}{\code{RandomVariable$normalizedFit()}}
\item \href{#method-RandomVariable-clone}{\code{RandomVariable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RandomVariable-density"></a>}}
\if{latex}{\out{\hypertarget{method-RandomVariable-density}{}}}
\subsection{Method \code{density()}}{
Provides the probability density for a given value in the 
   distribution of the random variable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandomVariable$density(val, log)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{The value for which the probability density is requested}

\item{\code{log}}{A boolean switch for requesting the log probability density}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The probability density of the value requested
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RandomVariable-markovStep"></a>}}
\if{latex}{\out{\hypertarget{method-RandomVariable-markovStep}{}}}
\subsection{Method \code{markovStep()}}{
Provides a new location in multivariate space based on a previous
   location provided as a vector of variable values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandomVariable$markovStep(vector, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vector}}{Previous location in multivariate space}

\item{\code{...}}{Generic arguments needed by internal functions}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Next location in multivariate space
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RandomVariable-normalizedFit"></a>}}
\if{latex}{\out{\hypertarget{method-RandomVariable-normalizedFit}{}}}
\subsection{Method \code{normalizedFit()}}{
Resets the variance statistics for the distribuiton that match
   the structure of provided data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandomVariable$normalizedFit(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Data to be fit}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The variance statistics providing the best fit of the 
   distribution to the data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RandomVariable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RandomVariable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandomVariable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
