% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomVariable.R
\name{RVMultivariateNormal}
\alias{RVMultivariateNormal}
\title{R6 class defining a multivariate normal distribution}
\description{
Provides tools for working with a multivariate random variable with
   a joint normal distribution defined by a covariance matrix
}
\section{Super class}{
\code{\link[inferno:RandomVariable]{inferno::RandomVariable}} -> \code{RVMultivariateNormal}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{means}}{Vector of means for the multivariate normal distribution}

\item{\code{covariances}}{Matrix of covariances for the multivariate normal distribution}

\item{\code{adjustCovarianceFactor}}{A scalar factor for adjusting the covariance in normalized fits}

\item{\code{tinyIdentMatrix}}{A identity matrix used to be sure normalized fits do not have
zero values}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RVMultivariateNormal-new}{\code{RVMultivariateNormal$new()}}
\item \href{#method-RVMultivariateNormal-markovStep}{\code{RVMultivariateNormal$markovStep()}}
\item \href{#method-RVMultivariateNormal-normalizedFit}{\code{RVMultivariateNormal$normalizedFit()}}
\item \href{#method-RVMultivariateNormal-clone}{\code{RVMultivariateNormal$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="inferno" data-topic="RandomVariable" data-id="density"><a href='../../inferno/html/RandomVariable.html#method-RandomVariable-density'><code>inferno::RandomVariable$density()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RVMultivariateNormal-new"></a>}}
\if{latex}{\out{\hypertarget{method-RVMultivariateNormal-new}{}}}
\subsection{Method \code{new()}}{
Construct a new instance of the class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RVMultivariateNormal$new(
  covariances,
  means = rep(0, nrow(covariances)),
  tinyIdentFactor = 1e-20,
  adjustCovarianceFactor = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{covariances}}{Covariance matrix of the variables defining the
distribution}

\item{\code{means}}{Optional means of the variables defining the distribution.
Default is all means are set to zero.}

\item{\code{tinyIdentFactor}}{A very small number used in a diagonal matrix to ensure
that no values on the diagonal of a fit covariance matrix
are zero. Defaults to 1E-20.}

\item{\code{adjustCovarianceFactor}}{Allows adjustment of a fit covariance matrix by a scalar
factor. Defaults to 1.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RVMultivariateNormal-markovStep"></a>}}
\if{latex}{\out{\hypertarget{method-RVMultivariateNormal-markovStep}{}}}
\subsection{Method \code{markovStep()}}{
Provides a new location in multivariate space based on a previous
   location provided as a vector of variable values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RVMultivariateNormal$markovStep(vector, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vector}}{Previous location in multivariate space}

\item{\code{...}}{Arguments passed to the mvrnorm function}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Next location in multivariate space
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RVMultivariateNormal-normalizedFit"></a>}}
\if{latex}{\out{\hypertarget{method-RVMultivariateNormal-normalizedFit}{}}}
\subsection{Method \code{normalizedFit()}}{
Resets the variance statistics for the distribuiton that match
   the structure of provided data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RVMultivariateNormal$normalizedFit(obsMatrix, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{obsMatrix}}{Data to be fit. Each column in the matrix represents a variable in the
multivariate distribution.  Each row represents an observation of the
variables.}

\item{\code{...}}{Arguments passed to the cov() function}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The variance statistics providing the best fit of the 
   distribution to the data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RVMultivariateNormal-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RVMultivariateNormal-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RVMultivariateNormal$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
