% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomVariable.R
\name{RVUniform}
\alias{RVUniform}
\title{R6 class defining a uniform random variable}
\description{
Provides tools for working with a random variable with
   a uniform distribution defined by minimum and maximum
   values
}
\section{Super class}{
\code{\link[inferno:RandomVariable]{inferno::RandomVariable}} -> \code{RVUniform}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{min}}{Minimum value}

\item{\code{max}}{Maximum value}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RVUniform$new()}}
\item \href{#method-density}{\code{RVUniform$density()}}
\item \href{#method-clone}{\code{RVUniform$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="inferno" data-topic="RandomVariable" data-id="markovStep">}\href{../../inferno/html/RandomVariable.html#method-markovStep}{\code{inferno::RandomVariable$markovStep()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="inferno" data-topic="RandomVariable" data-id="normalizedFit">}\href{../../inferno/html/RandomVariable.html#method-normalizedFit}{\code{inferno::RandomVariable$normalizedFit()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Construct a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RVUniform$new(min, max)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{min}}{Mininum value for the uniform distribution}

\item{\code{max}}{Maximum value for the uniform distribution}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-density"></a>}}
\if{latex}{\out{\hypertarget{method-density}{}}}
\subsection{Method \code{density()}}{
Method to provide the probability density for a given value in the 
   uniform distribution represented by the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RVUniform$density(val, log = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{val}}{The value for which the probability density is requested (numeric vector)}

\item{\code{log}}{A boolean switch for requesting the log probability density.
Defaults to FALSE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The probability density of the value requested (numeric vector).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RVUniform$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
