% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesFDistribution.R
\name{samplesFDistribution}
\alias{samplesFDistribution}
\title{Calculate posterior probabilities}
\usage{
samplesFDistribution(
  Y,
  X,
  learnt,
  subsamples,
  jacobian = TRUE,
  fn = NULL,
  combine = rbind,
  parallel = TRUE,
  silent = FALSE
)
}
\arguments{
\item{Y}{matrix or data.table: values of some variates of which we want
the joint probability one variate per column}

\item{X}{matrix or data.table: values of some variates conditional on
which we want the joint probability one variate per column}

\item{learnt}{Either a string with the name of a directory or full
path for an 'learnt.rds' object, or such an object itself}

\item{subsamples}{numeric: number of Monte Carlo samples to use}

\item{jacobian}{include the Jacobian in the output probability}

\item{fn}{NULL or function to apply to the group of MCsamples,
example 'identity' or 'mean'}

\item{combine}{how to combine the output for the different variate values}

\item{parallel}{logical or numeric: whether to use pre-existing parallel
workers, or how many to create and use}

\item{silent}{logical: give warnings or updates in the computation}
}
\value{
The frequencies F(Y|X) corresponding to the Monte Carlo samples
}
\description{
Calculate posterior probabilities
}
