% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulator.R
\name{Simulator}
\alias{Simulator}
\title{R6 class defining a simulator}
\description{
A simulator is designed to run a model based on a limited amount of
  model input that defines a modeling scenario. This limited input is
  characterized by a set of parameters for the model. The simulator also
  only returns the model output that is of interest in the scenario, and
  the output of interest is defined by the prediction extractor.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{parameterTranslator}}{The object used to translate parameters before running the model}

\item{\code{model}}{The model used to generate output}

\item{\code{predictionExtractor}}{The object used to extract a prediction from the model output}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Simulator-new}{\code{Simulator$new()}}
\item \href{#method-Simulator-setModel}{\code{Simulator$setModel()}}
\item \href{#method-Simulator-simulate}{\code{Simulator$simulate()}}
\item \href{#method-Simulator-clone}{\code{Simulator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulator-new"></a>}}
\if{latex}{\out{\hypertarget{method-Simulator-new}{}}}
\subsection{Method \code{new()}}{
Construct a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$new(parameterTranslator, model, predictionExtractor)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameterTranslator}}{A parameter translator object that is capable
of translating a simple vector of parameter values and inserting them
into the appropriate attributes of the model object, such that
the following execution of the model will generation a prediction
corresponding to those parameter values.}

\item{\code{model}}{The model used to simulate the scenario}

\item{\code{predictionExtractor}}{A prediction extractor object that is capable
of extracting a set of simple vectors from the model output, which can
then be compared to the observations in calculation of the objective
function value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulator-setModel"></a>}}
\if{latex}{\out{\hypertarget{method-Simulator-setModel}{}}}
\subsection{Method \code{setModel()}}{
Set the model used by the simulator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$setModel(model)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{The model used by the simulator to generate predictions
from a given parameter set}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulator-simulate"></a>}}
\if{latex}{\out{\hypertarget{method-Simulator-simulate}{}}}
\subsection{Method \code{simulate()}}{
Run a simulation based on the provided parameter values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$simulate(params = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{Optional set of parameters defining the scenario.
Defaults to NULL, which will run the model with its current input}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The prediction representing the output of interest in the scenario
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Simulator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Simulator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Simulator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
