% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayes.R
\name{StatsLoggerBayes}
\alias{StatsLoggerBayes}
\title{R6 class defining a statistics logger for a Bayesian objective function}
\description{
Extends \code{AdaptiveMCMCStatsLogger} to include stats specific to a Bayesian likelihood.
  See \code{\link{AdaptiveMCMCStatsLogger}} for documentation on inherited constructor.
}
\section{Super class}{
\code{\link[inferno:AdaptiveMCMCStatsLogger]{inferno::AdaptiveMCMCStatsLogger}} -> \code{StatsLoggerBayes}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{likelihoodColumn}}{Column index for the likelihood statistic}

\item{\code{propLikelihoodColumn}}{Column index for the proposed likelihood statistic}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-buildLog}{\code{StatsLoggerBayes$buildLog()}}
\item \href{#method-logAccepted}{\code{StatsLoggerBayes$logAccepted()}}
\item \href{#method-logRejected}{\code{StatsLoggerBayes$logRejected()}}
\item \href{#method-clone}{\code{StatsLoggerBayes$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="inferno" data-topic="AdaptiveMCMCStatsLogger" data-id="initialize">}\href{../../inferno/html/AdaptiveMCMCStatsLogger.html#method-initialize}{\code{inferno::AdaptiveMCMCStatsLogger$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="inferno" data-topic="AdaptiveMCMCStatsLogger" data-id="logFailedPredictions">}\href{../../inferno/html/AdaptiveMCMCStatsLogger.html#method-logFailedPredictions}{\code{inferno::AdaptiveMCMCStatsLogger$logFailedPredictions()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="inferno" data-topic="AdaptiveMCMCStatsLogger" data-id="writeFirstRow">}\href{../../inferno/html/AdaptiveMCMCStatsLogger.html#method-writeFirstRow}{\code{inferno::AdaptiveMCMCStatsLogger$writeFirstRow()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="inferno" data-topic="AdaptiveMCMCStatsLogger" data-id="writeRow">}\href{../../inferno/html/AdaptiveMCMCStatsLogger.html#method-writeRow}{\code{inferno::AdaptiveMCMCStatsLogger$writeRow()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-buildLog"></a>}}
\if{latex}{\out{\hypertarget{method-buildLog}{}}}
\subsection{Method \code{buildLog()}}{
Overrides the superclass buildLog method to add the structures needed
  for tracking Bayesian specific statistics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatsLoggerBayes$buildLog(numRows, objFunc, filePath)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{numRows}}{Number of rows of data expected in the log}

\item{\code{objFunc}}{The objective function with the statistics to be logged}

\item{\code{filePath}}{The file path where log files should be created}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logAccepted"></a>}}
\if{latex}{\out{\hypertarget{method-logAccepted}{}}}
\subsection{Method \code{logAccepted()}}{
Create accepted log entries
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatsLoggerBayes$logAccepted(index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{The row index for the stats to be logged}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-logRejected"></a>}}
\if{latex}{\out{\hypertarget{method-logRejected}{}}}
\subsection{Method \code{logRejected()}}{
Create rejected log entries
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatsLoggerBayes$logRejected(index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{The row index for the stats to be logged}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatsLoggerBayes$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
