% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetpr.R
\name{subset.probability}
\alias{subset.probability}
\title{Subset variates of an object of class \code{probability}}
\usage{
\method{subset}{probability}(x, subset)
}
\arguments{
\item{x}{Object of class "probability", obtained with \code{\link[=Pr]{Pr()}}.}

\item{subset}{Named list or named vector: variates to subset, given as list names, and corresponding values to subset.}
}
\value{
A list of class \code{probability}, identical to the original object \code{x} except for a reduced range of values in some if its variates.
}
\description{
An object of class \code{probability}, obtained with the \code{\link[=Pr]{Pr()}} function, holds the probabilities for all possible combinations of values of a set of joint variates \code{Y} conditional on a set of joint variates \code{X}, together with the variabilities of these probabilities and some other information. In some cases one may wish to exclude some of the values of the \code{Y} or \code{X} variates. For instance \code{Y} in the probability-class object could include the variate "age" with values from 18 to 100, and one may want to retain the values from 60 to 80.
}
\examples{
## Load the example `learnt` object included in the package
learnt <- learntExample

## Calculate the probability object for the three values of variate 'species',
## given values 43 and 44 of variate 'bill_len';
## this object contains probabilities, quantiles, and other information
probs <- Pr(
  Y = data.frame(species = c('Adelie', 'Chinstrap', 'Gentoo')),
  X = data.frame(bill_len = c(43, 44)),
  learnt = learnt, parallel = 1
)

probs$values

## Subset by retaining the values 'Adelie' and 'Gentoo' for species,
## and 44 for bill length
newprobs <- subset(
  probs,
  subset = list(species = c('Adelie', 'Gentoo'), bill_len = 43)
)

newprobs$values

## Plot these conditional probabilities and their variabilities
plot(newprobs)

hist(newprobs)

}
\seealso{
\code{\link[=Pr]{Pr()}}, which generates probability objects.

\code{\link[=plot.probability]{plot.probability()}} to plot probabilities and quantiles calculated by \code{Pr()} and subset by \code{\link[=subset.probability]{subset.probability()}}.

\code{\link[=hist.probability]{hist.probability()}} to plot histograms of the probability distributions calculated by \code{Pr()} and subset by \code{\link[=subset.probability]{subset.probability()}}.
}
