% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_tplotfunctions.R
\name{thist}
\alias{thist}
\title{Construct and plot histograms}
\usage{
thist(
  x,
  n = NULL,
  type = 6,
  pretty = FALSE,
  plot = FALSE,
  extendbreaks = FALSE,
  ...
)
}
\arguments{
\item{x}{Numeric or character: a collection of data or simulated values.}

\item{n}{integer or vector: number of bins to use, or boundaries of bins.}

\item{plot}{logical: plot the results? Default \code{FALSE}.}
}
\value{
A list of several elements. Most important: \code{counts}: the value count in each bin (frequency). \code{density}: the value count dividen by the bin width (frequency density). \code{mids}: the values of the bins' centres. \code{breaks}: the boundaries of the bins.
}
\description{
Utility function to construct and plot histograms, from samples obtained with \code{\link{Pr}}. The advantage of this function over \code{\link[base]{hist}} is that its \code{x} argument can be a vector of character values, as typically the case for nominal or ordinal variates (see \code{\link{metadata}}). The plot will have these characters as labels of the x-axis.
}
