% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_cleanup.R
\name{util_cleanup}
\alias{util_cleanup}
\title{Cleanup a learn()-output directory}
\usage{
util_cleanup(path)
}
\value{
No return value; called for side effects.
}
\description{
Cleanup a learn()-output directory
}
\details{
For deeper monitoring of the MCMC, the user can require the 'learn()' function not to clean intermediate MCMC-related files generated during the computation.

The present function can be used to remove these intermediate files from the output directory created by 'learn()'.
}
\keyword{internal}
