% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_lprob.R
\name{util_denorm}
\alias{util_denorm}
\title{Utility function to avoid finite-precision accuracys}
\usage{
util_denorm(lprob)
}
\description{
Utility function to avoid finite-precision accuracys
}
\keyword{internal}
