% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_cleanup.R
\name{util_joinPtraces}
\alias{util_joinPtraces}
\title{Join '____tempPtraces-' files}
\usage{
util_joinPtraces(path)
}
\description{
For deeper monitoring of the MCMC, the user can require the 'learn()' function not to clean intermediate MCMC-related files generated during the computation. The files with prefix '____tempPtraces-' contain chunks of MCMC traces.
}
\details{
The present function can be used to join them into a single trace.
}
\keyword{internal}
