% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_cleanup.R
\name{util_learntvar2sd}
\alias{util_learntvar2sd}
\title{Convert learnt with R/C/Dvar to learnt with R/C/Dsd}
\usage{
util_learntvar2sd(file)
}
\description{
Convert learnt with R/C/Dvar to learnt with R/C/Dsd
}
\keyword{internal}
