% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_lprob.R
\name{util_lprob}
\alias{util_lprob}
\title{Calculate collection of log-probabilities for different components and samples}
\usage{
util_lprob(
  x,
  learnt,
  nR,
  iR,
  tR,
  nC,
  iC,
  tC,
  Clefts,
  Crights,
  nD,
  iD,
  tD,
  Dsteps,
  Dlefts,
  Drights,
  nO,
  iO,
  tO,
  nN,
  iN,
  tN,
  nB,
  iB,
  tB
)
}
\arguments{
\item{nR}{etc: Parameters containing appropriate indices}

\item{X}{numerical matrix: transformed variates}

\item{learnt:}{Monte-Carlo-output object}
}
\value{
Matrix with as many rows as components and as many cols as samples
}
\description{
Calculate collection of log-probabilities for different components and samples
}
\keyword{internal}
