% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_lprobs.R
\name{util_lprobs}
\alias{util_lprobs}
\title{Calculate collection of log-probabilities for different components and samples}
\usage{
util_lprobs(
  x,
  learnt,
  nR,
  iR,
  tR,
  nLR,
  iLR,
  tLR,
  nUR,
  iUR,
  tUR,
  nC,
  iC,
  tC,
  nLC,
  iLC,
  tLC,
  nUC,
  iUC,
  tUC,
  Clefts,
  Crights,
  nD,
  iD,
  tD,
  nLD,
  iLD,
  tLD,
  nUD,
  iUD,
  tUD,
  Dsteps,
  Dlefts,
  Drights,
  nO,
  iO,
  tO,
  nLO,
  iLO,
  tLO,
  nUO,
  iUO,
  tUO,
  nN,
  iN,
  tN,
  nB,
  iB,
  tB
)
}
\arguments{
\item{nR}{etc: Parameters containing appropriate indices}

\item{X}{numerical matrix: transformed variates}

\item{learnt:}{Monte-Carlo-output object}
}
\value{
Matrix with as many rows as components and as many cols as samples
}
\description{
Calculate collection of log-probabilities for different components and samples
}
\keyword{internal}
