% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_lprobsargs.R
\name{util_lprobsargs}
\alias{util_lprobsargs}
\title{Prepare arguments for util_lprobs from data}
\usage{
util_lprobsargs(x, auxmetadata, learnt, cumul = NULL)
}
\description{
Prepare arguments for util_lprobs from data
}
\keyword{internal}
