% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_lprobsargsmi.R
\name{util_lprobsargsmi}
\alias{util_lprobsargsmi}
\title{Prepare arguments for util_lprobs from data}
\usage{
util_lprobsargsmi(x, auxmetadata, learnt)
}
\description{
Prepare arguments for util_lprobs from data
}
\keyword{internal}
