% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_lprobs.R
\name{util_lprobsargsyx}
\alias{util_lprobsargsyx}
\title{Prepare arguments for util_lprobsyx from data}
\usage{
util_lprobsargsyx(x, auxmetadata, learnt, tails = NULL)
}
\description{
Prepare arguments for util_lprobsyx from data
}
\keyword{internal}
