% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_lprobs.R
\name{util_lprobsbase}
\alias{util_lprobsbase}
\title{Calculate collection of log-probabilities for different components and samples}
\usage{
util_lprobsbase(xVs, params, logW, temporarydir = NULL, lab = "")
}
\value{
Matrix with as many rows as components and as many cols as samples
}
\description{
Calculate collection of log-probabilities for different components and samples
}
\keyword{internal}
