% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_lprob.R
\name{util_lprobsmi}
\alias{util_lprobsmi}
\title{Calculate pairs of log-probabilities for mutualinfo()}
\usage{
util_lprobsmi(xVs, params1, params2, lWnorm, lW)
}
\description{
Calculate pairs of log-probabilities for mutualinfo()
}
\keyword{internal}
