% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_checkpoints.R
\name{util_Pcheckpoints}
\alias{util_Pcheckpoints}
\title{Calculate joint frequencies for MCMC-monitoring checkpoints}
\usage{
util_Pcheckpoints(testdata, learnt)
}
\arguments{
\item{testdata}{List of objects calculated with util_prepPcheckpoints}

\item{learnt}{mcsamples object}
}
\value{
The joint frequencies of Y corresponding to the Monte Carlo samples
}
\description{
Used in 'learn()'.
}
\keyword{internal}
