% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_checkpoints.R
\name{util_prepPcheckpoints}
\alias{util_prepPcheckpoints}
\title{Calculate joint frequencies for checkpoints in learn()}
\usage{
util_prepPcheckpoints(x, auxmetadata, pointsid = NULL)
}
\value{
some arguments to be repeatedly used in util_Pcheckpoints
}
\description{
Calculate joint frequencies for checkpoints in learn()
}
\keyword{internal}
