% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_lprobs.R
\name{util_qYXcont}
\alias{util_qYXcont}
\title{Calculate quantiles for continuous Y by bisection}
\usage{
util_qYXcont(
  iyx,
  params1,
  params2,
  auxmetadata,
  temporarydir,
  usememory = TRUE,
  doquantiles,
  quantiles,
  dosamples,
  nsamples,
  Qerror,
  tol = .Machine$double.eps * 3
)
}
\description{
Calculate quantiles for continuous Y by bisection
}
\keyword{internal}
