% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_lprobs.R
\name{util_qYXdiscr}
\alias{util_qYXdiscr}
\title{Calculate quantiles for discrete Y by bisection}
\usage{
util_qYXdiscr(
  iyx,
  params1,
  params2,
  auxmetadata,
  temporarydir,
  usememory = TRUE,
  doquantiles,
  quantiles,
  dosamples,
  nsamples,
  Qerror = NULL,
  tol = NULL
)
}
\description{
Calculate quantiles for discrete Y by bisection
}
\keyword{internal}
