% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_vtransform.R
\name{vtransform}
\alias{vtransform}
\title{Transforms variates to different representations}
\usage{
vtransform(
  x,
  auxmetadata,
  Rout = NULL,
  Cout = NULL,
  Dout = NULL,
  Bout = NULL,
  Oout = NULL,
  Nout = NULL,
  variates = NULL,
  logjacobianOr = NULL
)
}
\arguments{
\item{x}{data.table object containing data to be transformed}

\item{auxmetadata}{auxmetadata object}

\item{Rout}{String, output of R-type variate, with possible values:
'normalized': for internal MCMC use
'mi': for use in mutualinfo()
'original': original representation}

\item{Cout}{String, output of C-type variate, with possible values:
'init': for internal MCMC use (init input)
'left', 'right': for internal MCMC use
'aux', 'lat': for internal MCMC use
'boundnormalized': for sampling functions
'boundisinf': for sampling functions
'mi': for use in mutualinfo()
'original': original representation}

\item{Dout}{String, output of D-type variate, with possible values:
'init': for internal MCMC use (init input)
'left', 'right': for internal MCMC use
'aux': for internal MCMC use
'boundisinf': for sampling functions
'normalized': for sampling functions
'mi': for use in mutualinfo()
'original': original representation}

\item{Bout}{String, output of B-type variate, with possible values:
'numeric': for internal MCMC use, values 0,1
'original': original representation}

\item{Oout}{String, output of O-type variate, with possible values:
'numeric': for internal MCMC use, values 1,2,...
'original': original representation}

\item{Nout}{String, output of N-type variate, with possible values:
'numeric': for internal MCMC use, values 1,2,...
'original': original representation}

\item{variates}{String vector, names of variates
corresponding to columns of x (in case x misses column names)}

\item{logjacobianOr}{Logical or \code{NULL}: output is the log-Jacobian in orginal or transformed domain? \code{NULL} (default) means do not calculate the log-Jacobians}
}
\value{
data frame of transformed variates, or their log-Jacobians
}
\description{
Transforms variates to different representations
}
\keyword{internal}
