% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdaptiveMCMCStatsLogger.R
\name{AdaptiveMCMCStatsLogger}
\alias{AdaptiveMCMCStatsLogger}
\title{R6 class defining a logger for an adaptive MCMC algorithm}
\description{
Logs the stats from a Markov Chain accept/reject algorithm. Stats
   are logged both to a matrix in memory and written to files, if configured
   to do so.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{numRows}}{Number of rows of data}

\item{\code{objFunc}}{Objective function used to calculate statistics}

\item{\code{stats}}{The data frame containing the statistics}

\item{\code{statsFile}}{The file used to track the statistics}

\item{\code{filePath}}{Path to the directory where the stats file should be written}

\item{\code{isLoggingFailedPredictions}}{Controls whether failed predictions are logged or not}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AdaptiveMCMCStatsLogger-new}{\code{AdaptiveMCMCStatsLogger$new()}}
\item \href{#method-AdaptiveMCMCStatsLogger-buildLog}{\code{AdaptiveMCMCStatsLogger$buildLog()}}
\item \href{#method-AdaptiveMCMCStatsLogger-logAccepted}{\code{AdaptiveMCMCStatsLogger$logAccepted()}}
\item \href{#method-AdaptiveMCMCStatsLogger-logRejected}{\code{AdaptiveMCMCStatsLogger$logRejected()}}
\item \href{#method-AdaptiveMCMCStatsLogger-logFailedPredictions}{\code{AdaptiveMCMCStatsLogger$logFailedPredictions()}}
\item \href{#method-AdaptiveMCMCStatsLogger-writeFirstRow}{\code{AdaptiveMCMCStatsLogger$writeFirstRow()}}
\item \href{#method-AdaptiveMCMCStatsLogger-writeRow}{\code{AdaptiveMCMCStatsLogger$writeRow()}}
\item \href{#method-AdaptiveMCMCStatsLogger-clone}{\code{AdaptiveMCMCStatsLogger$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdaptiveMCMCStatsLogger-new"></a>}}
\if{latex}{\out{\hypertarget{method-AdaptiveMCMCStatsLogger-new}{}}}
\subsection{Method \code{new()}}{
Construct a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdaptiveMCMCStatsLogger$new(
  filePath = NULL,
  statsFile = "stats.csv",
  isLoggingFailedPredictions = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filePath}}{Optional argument to set the path to which output files are written}

\item{\code{statsFile}}{Name of the file with logged stats data}

\item{\code{isLoggingFailedPredictions}}{Controls whether failed predictions are logged or not}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdaptiveMCMCStatsLogger-buildLog"></a>}}
\if{latex}{\out{\hypertarget{method-AdaptiveMCMCStatsLogger-buildLog}{}}}
\subsection{Method \code{buildLog()}}{
Creates the data structures and files necessary to track
   the statistics through a data analysis
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdaptiveMCMCStatsLogger$buildLog(numRows, objFunc, filePath = "./output")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{numRows}}{Number of rows of data expected in the log}

\item{\code{objFunc}}{The objective function with the statistics to be logged}

\item{\code{filePath}}{The file path where log files should be created}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdaptiveMCMCStatsLogger-logAccepted"></a>}}
\if{latex}{\out{\hypertarget{method-AdaptiveMCMCStatsLogger-logAccepted}{}}}
\subsection{Method \code{logAccepted()}}{
Create accepted log entries
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdaptiveMCMCStatsLogger$logAccepted(index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{Index of row for results}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdaptiveMCMCStatsLogger-logRejected"></a>}}
\if{latex}{\out{\hypertarget{method-AdaptiveMCMCStatsLogger-logRejected}{}}}
\subsection{Method \code{logRejected()}}{
Create rejected log entries
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdaptiveMCMCStatsLogger$logRejected(index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{Index of row for results}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdaptiveMCMCStatsLogger-logFailedPredictions"></a>}}
\if{latex}{\out{\hypertarget{method-AdaptiveMCMCStatsLogger-logFailedPredictions}{}}}
\subsection{Method \code{logFailedPredictions()}}{
Create failed prediction log entries
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdaptiveMCMCStatsLogger$logFailedPredictions(index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{Index of row for results}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdaptiveMCMCStatsLogger-writeFirstRow"></a>}}
\if{latex}{\out{\hypertarget{method-AdaptiveMCMCStatsLogger-writeFirstRow}{}}}
\subsection{Method \code{writeFirstRow()}}{
Write the first row of output files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdaptiveMCMCStatsLogger$writeFirstRow()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
No defined return value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdaptiveMCMCStatsLogger-writeRow"></a>}}
\if{latex}{\out{\hypertarget{method-AdaptiveMCMCStatsLogger-writeRow}{}}}
\subsection{Method \code{writeRow()}}{
Write a row to output files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdaptiveMCMCStatsLogger$writeRow(index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{Index of row for results}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AdaptiveMCMCStatsLogger-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AdaptiveMCMCStatsLogger-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AdaptiveMCMCStatsLogger$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
