% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayes.R
\name{BayesLogLikelihood}
\alias{BayesLogLikelihood}
\title{R6 class defining a basic Bayesian objective function}
\description{
Provides the tools for calculating a Bayes posterior 
  likelihood value for the comparison of a model prediction and
  observations. The Bayes likelihood is calculated by adding
  the summed log likelihood from prior distributions of
  parameters to the log likelihood of a provided objective 
  function.
}
\section{Super class}{
\code{\link[inferno:ObjectiveFunction]{inferno::ObjectiveFunction}} -> \code{BayesLogLikelihood}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{paramDists}}{A list of random variables representing the prior
prior probabilites for the parameters being estimated}

\item{\code{logLikelihood}}{The objective function that will calculate the 
fit metric (most likely a log likelihood) that will be added to the
sum of the prior log likelihoods to generate the overall objective
function value}

\item{\code{logPriors}}{The sum of the log of the prior probabilites}

\item{\code{negate}}{A flag that determines if the obective function should be negated
(likely for use with minimizing optimizers)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BayesLogLikelihood$new()}}
\item \href{#method-propose}{\code{BayesLogLikelihood$propose()}}
\item \href{#method-realize}{\code{BayesLogLikelihood$realize()}}
\item \href{#method-compare}{\code{BayesLogLikelihood$compare()}}
\item \href{#method-clone}{\code{BayesLogLikelihood$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="inferno" data-topic="ObjectiveFunction" data-id="plotFit">}\href{../../inferno/html/ObjectiveFunction.html#method-plotFit}{\code{inferno::ObjectiveFunction$plotFit()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Construct a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BayesLogLikelihood$new(paramDists, logLikelihood, negate = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{paramDists}}{A list of random variables representing the prior
prior probabilites for the parameters being estimated}

\item{\code{logLikelihood}}{The objective function that will calculate the 
fit metric (most like a log likelihood) that will be added to the
sum of the prior log likelihoods to generate the overall objective
function value}

\item{\code{negate}}{Optional switch to negate the objective function value
to adjust for algorthims that minimize or maximize objective funciton
values}

\item{\code{...}}{Additional required arguments for the constructor 
of the superclass \code{ObjectiveFunction} must be provided. Optional
arguments are also available for various features. 
See documentation for the class \code{\link{ObjectiveFunction}} for a description
of the additional required and optional arguments.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-propose"></a>}}
\if{latex}{\out{\hypertarget{method-propose}{}}}
\subsection{Method \code{propose()}}{
Overrides the superclass propose method to call the base objective 
  function propose method, before adding the log of the prior probabilities
  to produce the Bayesian posterior likelihood.

  Proposes a model with a given permutation represented by a subset of
  input
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BayesLogLikelihood$propose(params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{The subset of input being proposed as a permutation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The value of the objective function
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-realize"></a>}}
\if{latex}{\out{\hypertarget{method-realize}{}}}
\subsection{Method \code{realize()}}{
Overrides the superclass realize method to be sure realize is called
  for the base objective function and the observation attribute is set
  appropriately.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BayesLogLikelihood$realize()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The synthetic observation created
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compare"></a>}}
\if{latex}{\out{\hypertarget{method-compare}{}}}
\subsection{Method \code{compare()}}{
Compares a list of prediction vectors to an associated list of
  observation vectors based on a Bayesian posterior likelihood.

  This objective function uses a base objective function that to which 
  the logarithms of prior probabilities are added to for a Bayesian 
  objective context. For example, a formal Bayes likelihood based on
  independent and normally distributed error would use a formal log
  likelihood as the base objective function.

  Note that handling of NA values depend on the "ignore.na" attribute
  that is configurable in the constructor. By default NAs are ignored.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BayesLogLikelihood$compare(params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{Any "NaN" values in the sd attribute will be replaced by parameters
from the end of the parameter vector. (This feature is critical
if these standard deviations are being inferred as part of an
inferential modeling algorithm).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The log likelihood (will be negated if negate attribute is TRUE)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BayesLogLikelihood$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
