% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_buildauxmetadata.R
\name{buildauxmetadata}
\alias{buildauxmetadata}
\title{Build preliminary metadata flie}
\usage{
buildauxmetadata(data, metadata, Dthreshold = 1)
}
\arguments{
\item{data}{data.frame object}

\item{metadata}{data.frame object}

\item{Dthreshold}{positive number: threshold of fraction
of unique datapoints to total datapoints, to decide
whether to treat a rounded variate as continuous}
}
\value{
an auxmetadata data.frame object
}
\description{
Build preliminary metadata flie
}
\keyword{internal}
