% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmetadata.R
\name{buildmetadata}
\alias{buildmetadata}
\alias{metadata}
\title{Metadata}
\usage{
buildmetadata(
  data,
  file = NULL,
  includevrt = NULL,
  excludevrt = NULL,
  addsummary2metadata = FALSE,
  backupfiles = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{A dataset, given as a \code{\link[base]{data.frame}}
or as a file path to a csv file.}

\item{file}{String: name of csv file where the metadata should be saved;
if \code{NULL}: output metadata as \code{VALUE}.}

\item{includevrt}{Character or \code{NULL}: name of variates in dataset to be included.}

\item{excludevrt}{Character or \code{NULL}: name of variates in dataset to be excluded.}

\item{addsummary2metadata}{Logical: also output some diagnostic statistics
in the metadata? Default \code{FALSE}.}

\item{backupfiles}{Logical: rename previous metadata file if it exists?
Default \code{TRUE}.}

\item{verbose}{Logical: output heuristics for each variate? Default \code{TRUE}.}
}
\value{
If \code{file = NULL}, a preliminary metadata file is created
and \code{VALUE} is \code{NULL};
otherwise \code{VALUE} is a \code{\link[base]{data.frame}} containing the metadata.
}
\description{
Details about metadata, metadata file and object, and utility function to build a metadata file or object.
}
\details{
In order to correctly learn from a dataset, the \code{\link{learn}} function needs information that is not contained in the data themeselves; that is, it needs \emph{meta}data. Metadata are provided either as a \code{csv} file or as a \code{\link[base]{data.frame}}.

A metadata file must contain the following fields, even if some of them may be empty:

\code{name}, \code{type}, \code{Nvalues}, \code{datastep}, \code{domainmin}, \code{domainmax}, \code{minincluded}, \code{maxinluded}, \code{plotmin}, \code{plotmax}, \code{V1}, \code{V2}, ...

with possibly additional \code{V}-fields, sequentially numbered.

The \code{type} field has four possible values:

\code{binary}, \code{nominal}, \code{ordinal}, \code{continuous}

Which fields cannot be empty depend on the \code{type} field. Here is a list of require
}
\section{Necessity of metadata}{


This is a test
}

