% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_mcmcstop.R
\name{funAC}
\alias{funAC}
\title{Compute autocovariance}
\usage{
funAC(y)
}
\arguments{
\item{y}{Time series}
}
\value{
Autocovariances at different lags
}
\description{
From \strong{rstan} \url{https://github.com/stan-dev/rstan/blob/develop/rstan/rstan/R/monitor.R}
}
\keyword{internal}
