% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_mcmcstop.R
\name{funESS3}
\alias{funESS3}
\title{Compute ESS}
\usage{
funESS3(x)
}
\arguments{
\item{x}{Vector of Monte Carlo samples}
}
\value{
Effective Sample Size
}
\description{
From \strong{rstan} \url{https://github.com/stan-dev/rstan/blob/develop/rstan/rstan/R/monitor.R}
}
\keyword{internal}
