% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_mcmcstop.R
\name{funMCEQ}
\alias{funMCEQ}
\title{Calculate credibility quantiles on estimated quantile}
\usage{
funMCEQ(x, prob = c(0.055, 0.945), Qpair = pnorm(c(-1, 1)))
}
\arguments{
\item{x:}{a vector of MC samples}

\item{prob:}{quantile whose error intervalis being estimated}

\item{Qpair:}{lower and higher credibility-quantiles requested}
}
\value{
Estimates lower and higher credibility-quantiles on estimated quantile
}
\description{
Modified from Vehtari et al.
}
\keyword{internal}
