% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_tplotfunctions.R
\name{hist.probability}
\alias{hist.probability}
\title{Plot the variability of an object of class "probability" as a histogram}
\usage{
\method{hist}{probability}(
  p,
  breaks = NULL,
  legend = TRUE,
  lty = c(1, 2, 4, 3, 6, 5),
  lwd = 2,
  col = palette(),
  alpha.f = 0.125,
  xlab = NULL,
  ylab = NULL,
  xlim = NULL,
  ylim = c(0, NA),
  grid = TRUE,
  add = FALSE,
  ...
)
}
\arguments{
\item{p}{object of class "probability", obtained with \code{\link{Pr}} or \code{\link{tailPr}}.}

\item{breaks}{\code{NULL} or as in function \code{\link[graphics]{hist}}. If \code{NULL} (default), an optimal number of breaks for each probability distribution is computed.}

\item{legend}{string or logical: plot a legend of the different curves at position \code{legend}? If \code{TRUE}, position is 'top'.}

\item{alpha.f}{Numeric, default 0.125: opacity of the histogram filling. \code{0} means no filling.}

\item{...}{other parameters to be passed to \code{\link{flexiplot}}.}
}
\description{
This \code{\link[graphics]{hist}}ogram method is a utility to visualize the variability of the probabilities obtained with \code{\link{Pr}} or \code{\link{tailPr}}, which can also be interpreted as the probability density for the whole-population frequencies.
}
