% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_tplotfunctions.R
\name{hist.probability}
\alias{hist.probability}
\title{Plot the variability of an object of class "probability" as a histogram}
\usage{
\method{hist}{probability}(
  p,
  breaks = NULL,
  legend = "top",
  lty = c(1, 2, 4, 3, 6, 5),
  lwd = 2,
  col = palette(),
  alpha.f = 1,
  fill.alpha.f = 0.125,
  showmean = TRUE,
  xlab = NULL,
  ylab = NULL,
  xlim = NULL,
  ylim = c(0, NA),
  main = NULL,
  grid = TRUE,
  add = FALSE,
  ...
)
}
\arguments{
\item{p}{Object of class "probability", obtained with \code{\link[=Pr]{Pr()}} or \code{\link[=tailPr]{tailPr()}}.}

\item{breaks}{\code{NULL} or as in function \code{\link[graphics:hist]{graphics::hist()}}. If \code{NULL} (default), an optimal number of breaks for each probability distribution is computed.}

\item{legend}{One of the values \code{"bottomright"}, \code{"bottom"}, \code{"bottomleft"}, \code{"left"}, \code{"topleft"}, \code{"top"}, \code{"topright"}, \code{"right"}, \code{"center"} (see \code{\link[graphics:legend]{graphics::legend()}}): plot a legend at that position. A value \code{FALSE} or any other does not plot any legend. Default \code{"top"}.}

\item{fill.alpha.f}{Numeric, default 0.125: opacity of the histogram filling. \code{0} means no filling.}

\item{showmean}{Logical, default \code{TRUE}: show the means of the probability distributions? The means correspond to the probabilities about the next observed unit.}

\item{...}{Other parameters to be passed to \code{\link[=flexiplot]{flexiplot()}}.}
}
\description{
This \code{\link[graphics:hist]{graphics::hist()}}ogram method is a utility to visualize the variability of the probabilities obtained with \code{\link[=Pr]{Pr()}} or \code{\link[=tailPr]{tailPr()}}, which can also be interpreted as the probability density for the whole-population frequencies.
}
