% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferno-package.R
\docType{package}
\name{inferno-package}
\alias{inferno}
\alias{inferno-package}
\title{inferno: Bayesian nonparametric exchangeable inference - model-free uncertainty-quantified prediction}
\description{
Offers several functions for Bayesian nonparametric density/population/exchangeable inference, including Monte Carlo calculation of posterior densities; from a machine-learning perspective, it offers a model-free, uncertainty-quantified predction algorithm.
}
\author{
\strong{Maintainer}: PierGianLuca Porta Mana \email{pgl@portamana.org} (\href{https://orcid.org/0000-0002-6070-0784}{ORCID})

Authors:
\itemize{
  \item Aurora S. Grefsrud \email{Aurora.Singstad.Grefsrud@hvl.no} (\href{https://orcid.org/0009-0005-9063-4131}{ORCID})
}

}
\keyword{internal}
