% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferno-package.R
\docType{package}
\name{inferno-package}
\alias{inferno}
\alias{inferno-package}
\title{inferno: Bayesian nonparametric population inference}
\description{
Functions for Bayesian nonparametric population (or exchangeable, or density) inference. From a machine-learning perspective, it offers a model-free, uncertainty-quantified prediction algorithm.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pglpm/inferno/}
  \item \url{https://pglpm.github.io/inferno/}
}

}
\author{
\strong{Maintainer}: PierGianLuca Porta Mana \email{pgl@portamana.org} (\href{https://orcid.org/0000-0002-6070-0784}{ORCID})

Authors:
\itemize{
  \item Aurora Grefsrud \email{agre@hvl.no} (\href{https://orcid.org/0009-0005-9063-4131}{ORCID})
  \item Håkon Mydland \email{haakon.mydland@gmail.com} (\href{https://orcid.org/0009-0004-9558-8894}{ORCID})
  \item Maksim Ohvrill \email{maksimohvrill@hotmail.com}
  \item Simen Hesthamar Hauge \email{simen@hhnet.no} (\href{https://orcid.org/0009-0005-6158-2846}{ORCID})
}

}
\keyword{internal}
