% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferno-package.R
\docType{package}
\name{inferno-package}
\alias{inferno}
\alias{inferno-package}
\title{inferno: Bayesian nonparametric exchangeable inference - model-free uncertainty-quantified prediction}
\description{
Offers several functions for Bayesian nonparametric exchangeable inference (also called "density" or "population" inference), including Monte Carlo calculation of posterior densities. From a machine-learning perspective, it offers a model-free, uncertainty-quantified predction algorithm.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pglpm/inferno/}
  \item \url{https://pglpm.github.io/inferno/}
}

}
\author{
\strong{Maintainer}: PierGianLuca Porta Mana \email{pgl@portamana.org} (\href{https://orcid.org/0000-0002-6070-0784}{ORCID})

Authors:
\itemize{
  \item Aurora Grefsrud \email{agre@hvl.no} (\href{https://orcid.org/0009-0005-9063-4131}{ORCID})
}

}
\keyword{internal}
