% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_inits_inferpopulation.R,
%   R/_newinferpopulation.R, R/inferpopulation.R
\name{inferpopulation}
\alias{inferpopulation}
\title{Monte Carlo computation of posterior distribution of population frequencies}
\usage{
inferpopulation(
  data,
  metadata,
  auxdata = NULL,
  outputdir,
  nsamples = 3600,
  nchains = 60,
  nsamplesperchain = 60,
  parallel = TRUE,
  seed = NULL,
  cleanup = TRUE,
  appendtimestamp = TRUE,
  appendinfo = TRUE,
  output = "directory",
  subsampledata = NULL,
  startupMCiterations = 3600,
  minMCiterations = 0,
  maxMCiterations = +Inf,
  maxhours = +Inf,
  ncheckpoints = NULL,
  relerror = 0.05,
  prior = missing(data),
  thinning = NULL,
  plottraces = TRUE,
  showKtraces = FALSE,
  showAlphatraces = FALSE,
  hyperparams = list(nclusters = 64, minalpha = -4, maxalpha = 4, byalpha = 1, Rshapelo =
    0.5, Rshapehi = 0.5, Rvarm1 = 3^2, Cshapelo = 0.5, Cshapehi = 0.5, Cvarm1 = 3^2,
    Dshapelo = 0.5, Dshapehi = 0.5, Dvarm1 = 3^2, Lshapelo = 0.5, Lshapehi = 0.5, Lvarm1
    = 3^2, Bshapelo = 1, Bshapehi = 1, Dthreshold = 1)
)

inferpopulation(
  data,
  metadata,
  auxdata = NULL,
  outputdir,
  nsamples = 3600,
  nchains = 60,
  nsamplesperchain = 60,
  parallel = TRUE,
  seed = NULL,
  cleanup = TRUE,
  appendtimestamp = TRUE,
  appendinfo = TRUE,
  output = "directory",
  subsampledata = NULL,
  startupMCiterations = 3600,
  minMCiterations = 0,
  maxMCiterations = +Inf,
  maxhours = +Inf,
  ncheckpoints = NULL,
  relerror = 0.05,
  prior = missing(data),
  thinning = NULL,
  plottraces = TRUE,
  showKtraces = FALSE,
  showAlphatraces = FALSE,
  hyperparams = list(nclusters = 64, minalpha = -4, maxalpha = 4, byalpha = 1, Rshapelo =
    0.5, Rshapehi = 0.5, Rvarm1 = 3^2, Cshapelo = 0.5, Cshapehi = 0.5, Cvarm1 = 3^2,
    Dshapelo = 0.5, Dshapehi = 0.5, Dvarm1 = 3^2, Lshapelo = 0.5, Lshapehi = 0.5, Lvarm1
    = 3^2, Bshapelo = 1, Bshapehi = 1, Dthreshold = 1)
)

inferpopulation(
  data,
  metadata,
  auxdata = NULL,
  outputdir,
  nsamples = 3600,
  nchains = 60,
  nsamplesperchain = 60,
  parallel = TRUE,
  seed = NULL,
  cleanup = TRUE,
  appendtimestamp = TRUE,
  appendinfo = TRUE,
  output = "directory",
  subsampledata = NULL,
  startupMCiterations = 3600,
  minMCiterations = 0,
  maxMCiterations = +Inf,
  maxhours = +Inf,
  ncheckpoints = NULL,
  relerror = 0.05,
  prior = missing(data),
  thinning = NULL,
  plottraces = TRUE,
  showKtraces = FALSE,
  showAlphatraces = FALSE,
  hyperparams = list(nclusters = 64, minalpha = -4, maxalpha = 4, byalpha = 1, Rshapelo =
    0.5, Rshapehi = 0.5, Rvarm1 = 3^2, Cshapelo = 0.5, Cshapehi = 0.5, Cvarm1 = 3^2,
    Dshapelo = 0.5, Dshapehi = 0.5, Dvarm1 = 3^2, Lshapelo = 0.5, Lshapehi = 0.5, Lvarm1
    = 3^2, Bshapelo = 1, Bshapehi = 1, Dthreshold = 1)
)
}
\arguments{
\item{data}{data.frame object or filepath: datapoints}

\item{metadata}{Either the name of the csv file containing metadata
of the current dataset, or a data.frame with the metadata}

\item{auxdata}{NULL or data.frame object or filepath: extra datapoints
which would be too many to use in the Monte Carlo sampling,
but can be used to calculate hyperparameters}

\item{outputdir}{String, path to output file folder ## Rename to
outputPrefix, also addSuffix?}

\item{nsamples}{Integer, nr of desired MC samples}

\item{nchains}{Integer, nr of MC chains}

\item{nsamplesperchain}{Integer, nr of MC samples per chain}

\item{parallel, }{logical or numeric: whether to use pre-existing parallel
workers, or how many to create and use}

\item{seed}{Integer: seed for random number generator. If left as default
NULL, a random seed based on the system clock is used in the
set.seed() function}

\item{cleanup}{logical, default TRUE, removes files that can be used for
debugging}

\item{appendtimestamp}{logical, default TRUE: add timestamp to name of
output directory}

\item{appendinfo}{logical, default TRUE: add some simulation information
to name of output directory}

\item{output}{if string 'directory', return the output directory name;
if string 'mcoutput', return the 'Fdistribution' object;
anything else, no output}

\item{subsampledata}{Numeric: use only a subsample of the datapoints in 'data'}

\item{startupMCiterations}{Number of initial (burn-in) MC iterations}

\item{minMCiterations}{Minimum number of MC iterations to be done}

\item{maxMCiterations}{Maximum number of MC iterations}

\item{maxhours}{Maximum (approximately) time in hours for MC computation}

\item{ncheckpoints}{NULL (default), positive integer, or Inf:
number of datapoints to use for stopping the sampling;
if NULL, equal to number of variates + 2; if Inf, use all datapoints}

\item{relerror}{positive real: relative error of calculated probabilities
with respect to their variability with new data. It is only approximate}

\item{prior}{logical: Calculate the prior distribution of F?}

\item{thinning}{If NULL, let the diagnostics decide the MC thinning;
if positive, use this thinning value}

\item{plottraces}{logical: plot MC traces of diagnostic values}

\item{showKtraces}{logical, when true, it saves the K parameter during
sampling and plots its trace and histogram at the end. Keeping it to
FALSE (default) saves a little computation time.}

\item{showAlphatraces}{logical, : when true, it saves the Alpha parameter
more frequently during sampling and plots its trace and histogram at
the end. Keeping it to FALSE (default) saves a little computation
time.}

\item{hyperparams}{list of hyperparameters (used for debugging)}
}
\value{
name of directory containing output files, or Fdistribution object,
or empty

name of directory containing output files, or Fdistribution object,
or empty

name of directory containing output files, or Fdistribution object,
or empty
}
\description{
Monte Carlo computation of posterior distribution of population frequencies

Monte Carlo computation of posterior distribution of population frequencies

Monte Carlo computation of posterior distribution of population frequencies
}
