% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn}
\alias{learn}
\title{Monte Carlo computation of posterior probability distribution conditional on given data}
\usage{
learn(
  data,
  metadata,
  auxdata = NULL,
  outputdir = NULL,
  nsamples = 3600,
  nchains = 60,
  nsamplesperchain = 60,
  parallel = TRUE,
  seed = NULL,
  cleanup = TRUE,
  appendtimestamp = TRUE,
  appendinfo = TRUE,
  output = "directory",
  subsampledata = NULL,
  startupMCiterations = 3600,
  minMCiterations = 0,
  maxMCiterations = +Inf,
  maxhours = +Inf,
  ncheckpoints = NULL,
  relerror = 0.05,
  prior = missing(data),
  thinning = NULL,
  plottraces = TRUE,
  showKtraces = FALSE,
  showAlphatraces = FALSE,
  hyperparams = list(ncomponents = 64, minalpha = -4, maxalpha = 4, byalpha = 1, Rshapelo
    = 0.5, Rshapehi = 0.5, Rvarm1 = 3^2, Cshapelo = 0.5, Cshapehi = 0.5, Cvarm1 = 3^2,
    Dshapelo = 0.5, Dshapehi = 0.5, Dvarm1 = 3^2, Lshapelo = 0.5, Lshapehi = 0.5, Lvarm1
    = 3^2, Bshapelo = 1, Bshapehi = 1, Dthreshold = 1, tscalefactor = 2)
)
}
\arguments{
\item{data}{A dataset, given as a \code{\link[base]{data.frame}}
or as a file path to a csv file.}

\item{metadata}{A metadata object, given either as a data.frame object,
or as a file pa to a csv file.}

\item{auxdata}{A larger dataset, given as a data.frame
or as a file path to a csv file. Such a dataset
would be too many to use in the Monte Carlo sampling,
but can be used to calculate hyperparameters.}

\item{outputdir}{Character: path to folder where the output should be saved. If omitted, a directory is created that has the same name as the data file but with suffix "\verb{_output_}".}

\item{nsamples}{Integer: number of desired Monte Carlo samples. Default 3600.}

\item{nchains}{Integer: number of Monte Carlo chains. Default 60.}

\item{nsamplesperchain}{Integer: number of Monte Carlo samples per chain.}

\item{parallel}{Logical: use pre-existing parallel workers from package
\code{\link[doParallel]{doParallel}}.
Or integer: create and use that many parallel workers. Default \code{TRUE}.}

\item{seed}{Integer: use this seed for the random number generator.
If missing or \code{NULL} (default), do not set the seed.}

\item{cleanup}{Logical: remove diagnostic files at the end of the computation?
Default \code{TRUE}.}

\item{appendtimestamp}{Logical: append a timestamp to the name of
the output directory \code{outputdir}? Default \code{TRUE}.}

\item{appendinfo}{Logical: append information about dataset and Monte Carlo
parameters to the name of the output directory \code{outputdir}? Default \code{TRUE}.}

\item{output}{Character: if \code{'directory'}, return the output directory name
as \code{VALUE}; if string \code{'learnt'}, return the \code{'learnt'} object
containing the parameters obtained from the Monte Carlo computation.
Any other value: \code{VALUE} is \code{NULL}.}

\item{subsampledata}{Integer: use only a subset of this many datapoints for
the Monte Carlo computation.}

\item{startupMCiterations}{Integer: number of initial (burn-in)
Monte Carlo iterations. Default 3600.}

\item{minMCiterations}{Integer: minimum number of Monte Carlo iterations
to be done. Default 0.}

\item{maxMCiterations}{Integer: Do at most this many Monte Carlo iterations.
Default \code{Inf}.}

\item{maxhours}{Numeric: approximate time limit, in hours, for the
Monte Carlo computation to last. Default \code{Inf}.}

\item{ncheckpoints}{Integer: number of datapoints to use
for checking when the Monte Carlo computation should end.
If NULL (default), this is equal to number of variates + 2.
If Inf, use all datapoints.}

\item{relerror}{Numeric: desired maximal relative error of calculated probabilities
with respect to their variability with new data.}

\item{prior}{Logical: Calculate the prior distribution?}

\item{thinning}{Integer: thin out the Monte Carlo samples by this value.
If NULL (default): let the diagnostics decide the thinning value.}

\item{plottraces}{Logical: save plots of the Monte Carlo traces
of diagnostic values? Default \code{TRUE}.}

\item{showKtraces}{Logical: save plots of the Monte Carlo traces
of the K parameter? Default \code{FALSE}.}

\item{showAlphatraces}{Logical: save plots of the Monte Carlo traces
of the Alpha parameter? Default \code{FALSE}.}

\item{hyperparams}{List: hyperparameters of the prior.}
}
\value{
Name of directory containing output files, or learnt object,
or \code{NULL}, depending on argument \code{output}.
}
\description{
Monte Carlo computation of posterior probability distribution conditional on given data
}
