% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihood.R
\name{LogLikelihood}
\alias{LogLikelihood}
\title{A log likelihood objective function class (R6 class generator)}
\description{
The R6 class generator for a log likelihood objective function
  object used by optimization algorithms. 
  For example, an instance of this class would be the objective function 
  used for in Maximum Likelihood Estimate (MLE) analysis.
}
\section{Super class}{
\code{\link[inferno:ObjectiveFunction]{inferno::ObjectiveFunction}} -> \code{LogLikelihood}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{sd}}{Standard deviation used for calculating likelihood}

\item{\code{negate}}{Flag determining if likelihood should be negated
(usually set to TRUE for minimizing optimizers)}

\item{\code{ignore.na}}{Flag determining if NA values are ignored}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LogLikelihood$new()}}
\item \href{#method-compare}{\code{LogLikelihood$compare()}}
\item \href{#method-clone}{\code{LogLikelihood$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="inferno" data-topic="ObjectiveFunction" data-id="plotFit">}\href{../../inferno/html/ObjectiveFunction.html#method-plotFit}{\code{inferno::ObjectiveFunction$plotFit()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="inferno" data-topic="ObjectiveFunction" data-id="propose">}\href{../../inferno/html/ObjectiveFunction.html#method-propose}{\code{inferno::ObjectiveFunction$propose()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="inferno" data-topic="ObjectiveFunction" data-id="realize">}\href{../../inferno/html/ObjectiveFunction.html#method-realize}{\code{inferno::ObjectiveFunction$realize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="inferno" data-topic="ObjectiveFunction" data-id="setModel">}\href{../../inferno/html/ObjectiveFunction.html#method-setModel}{\code{inferno::ObjectiveFunction$setModel()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="inferno" data-topic="ObjectiveFunction" data-id="setObservation">}\href{../../inferno/html/ObjectiveFunction.html#method-setObservation}{\code{inferno::ObjectiveFunction$setObservation()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Constructs a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogLikelihood$new(..., sd, negate = FALSE, ignore.na = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional required arguments for the constructor 
of the superclass \code{ObjectiveFunction} must be provided. Optional
arguments are also available for various features. 
See documentation for the class \code{\link{ObjectiveFunction}} for a description
of the additional required and optional arguments.}

\item{\code{sd}}{A vector of standard deviations to be used for calculating the 
likelihood. This vector shoud have the an element with a standard deviation 
corresponding to each of the columns in the dataframes containing the 
predictons and observations that are compared by the objective function.}

\item{\code{negate}}{A boolean switch indicating if the objective function value
should be negated for switching between maximization or minimization
algorithms. Note that this is a log likelihood, so negation is 
equivalent to inverting the probability representing the raw likelihood.
Default value if FALSE (log likelihood is not negated).}

\item{\code{ignore.na}}{Determines if NA elements will be ignored in objective function value calculation.
Default is TRUE (ignore NAs in calculation).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compare"></a>}}
\if{latex}{\out{\hypertarget{method-compare}{}}}
\subsection{Method \code{compare()}}{
Compares a list of prediction vectors to an associated list of
  observation vectors based on a formal log likelihood function
  and assumption of a normal distribution of residual error.

  Note that handling of NA values depend on the "ignore.na" attribute
  that is configurable in the constructor. By default NAs are ignored.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogLikelihood$compare(params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{Any "NaN" values in the sd attribute will be replaced by parameters values
from the end of the parameter vector. This feature is critical
if these standard deviations are being inferred as part of an
inferential modeling algorithm, rather than being provided as an
attribute of the object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The log likelihood value from the comparison
   (will be negated if negate attribute of the object is TRUE)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogLikelihood$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
