% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikelihoodAR1.R
\name{LogLikelihoodAR1}
\alias{LogLikelihoodAR1}
\title{R6 class defining an AR1 log likelihood objective function}
\description{
An objective function based on a log likelihood considering a
  first-order autocorrelation in the residual error.
}
\section{Super classes}{
\code{\link[inferno:ObjectiveFunction]{inferno::ObjectiveFunction}} -> \code{\link[inferno:LogLikelihood]{inferno::LogLikelihood}} -> \code{LogLikelihoodAR1}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{arCoeff}}{The first-order autoregression coefficient}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LogLikelihoodAR1-new}{\code{LogLikelihoodAR1$new()}}
\item \href{#method-LogLikelihoodAR1-compare}{\code{LogLikelihoodAR1$compare()}}
\item \href{#method-LogLikelihoodAR1-clone}{\code{LogLikelihoodAR1$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="inferno" data-topic="ObjectiveFunction" data-id="plotFit"><a href='../../inferno/html/ObjectiveFunction.html#method-ObjectiveFunction-plotFit'><code>inferno::ObjectiveFunction$plotFit()</code></a></li>
<li><span class="pkg-link" data-pkg="inferno" data-topic="ObjectiveFunction" data-id="propose"><a href='../../inferno/html/ObjectiveFunction.html#method-ObjectiveFunction-propose'><code>inferno::ObjectiveFunction$propose()</code></a></li>
<li><span class="pkg-link" data-pkg="inferno" data-topic="ObjectiveFunction" data-id="realize"><a href='../../inferno/html/ObjectiveFunction.html#method-ObjectiveFunction-realize'><code>inferno::ObjectiveFunction$realize()</code></a></li>
<li><span class="pkg-link" data-pkg="inferno" data-topic="ObjectiveFunction" data-id="setModel"><a href='../../inferno/html/ObjectiveFunction.html#method-ObjectiveFunction-setModel'><code>inferno::ObjectiveFunction$setModel()</code></a></li>
<li><span class="pkg-link" data-pkg="inferno" data-topic="ObjectiveFunction" data-id="setObservation"><a href='../../inferno/html/ObjectiveFunction.html#method-ObjectiveFunction-setObservation'><code>inferno::ObjectiveFunction$setObservation()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogLikelihoodAR1-new"></a>}}
\if{latex}{\out{\hypertarget{method-LogLikelihoodAR1-new}{}}}
\subsection{Method \code{new()}}{
Construct a new instance of the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogLikelihoodAR1$new(..., arCoeff)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments to pass to the constructor of the super class}

\item{\code{arCoeff}}{The autoregression coefficient}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogLikelihoodAR1-compare"></a>}}
\if{latex}{\out{\hypertarget{method-LogLikelihoodAR1-compare}{}}}
\subsection{Method \code{compare()}}{
Overrides the compare method in the super class to add consideration
  of autocorrelation.

  Compares a list of prediction vectors to an associated list of
  observation vectors based on a formal log likelihood function
  and assumption of a normal distribution of residual error with
  first-order autocorrelation.

  Note that handling of NA values depend on the "ignore.na" attribute
  that is configurable in the constructor. By default NAs are ignored.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogLikelihoodAR1$compare(params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{params}}{Vector of values for parameters being estimated}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The log likelihood with consideration of first-order autocorrelation
  in error (will be negated if negate attribute is TRUE)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LogLikelihoodAR1-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LogLikelihoodAR1-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogLikelihoodAR1$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
