% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutualinfo.R
\name{mutualinfo}
\alias{mutualinfo}
\title{Calculate mutual information between groups of joint variates}
\usage{
mutualinfo(
  Yvrt,
  Xvrt,
  mcoutput,
  nsamples = 3600,
  unit = "Sh",
  useOquantiles = FALSE,
  parallel = TRUE,
  silent = FALSE
)
}
\arguments{
\item{Yvrt}{String vector: first group of joint variates}

\item{Xvrt}{String vector: second group of joint variates}

\item{mcoutput}{Either a string with the name of a directory or full path
for a 'FDistribution.rds' object, or such an object itself}

\item{nsamples}{numeric: number of samples from which to approximately
calculate the mutual information. Default 3600}

\item{unit}{Either one of 'Sh' (default), 'Hart', 'nat', or a positive real}

\item{useOquantiles}{Bool: internal, use metadata quantiles for ordinal variates}

\item{parallel, }{Bool or numeric: whether to use pre-existing parallel
workers, or how many to create and use}

\item{silent}{Bool: give warnings or updates in the computation}
}
\value{
A list with the mutual information, its error, and its unit
}
\description{
Calculate mutual information between groups of joint variates
}
