% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutualinfo.R
\name{mutualinfo}
\alias{mutualinfo}
\title{Calculate mutual information between groups of joint variates}
\usage{
mutualinfo(
  Y1names,
  Y2names,
  X = NULL,
  learnt,
  nsamples = 3600,
  unit = "Sh",
  parallel = TRUE,
  silent = TRUE
)
}
\arguments{
\item{Y1names}{String vector: first group of joint variates}

\item{Y2names}{String vector or NULL: second group of joint variates}

\item{X}{matrix or data.frame or NULL: values of some variates conditional on
which we want the probabilities}

\item{learnt}{Either a string with the name of a directory or full path
for an 'learnt.rds' object, or such an object itself}

\item{nsamples}{numeric: number of samples from which to approximately
calculate the mutual information. Default 3600}

\item{unit}{Either one of 'Sh' (default), 'Hart', 'nat', or a positive real}

\item{parallel, }{logical or numeric: whether to use pre-existing parallel
workers, or how many to create and use}

\item{silent}{logical: give warnings or updates in the computation}
}
\value{
A list with the mutual information, its error, and its unit
}
\description{
Calculate mutual information between groups of joint variates
}
