% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObservationGenerator.R
\name{ObservationGenerator}
\alias{ObservationGenerator}
\title{R6 class defining an artificial observation generator}
\description{
Generates an artificial observation data set base on a model.

  This class is abstract and is not intended to be instantiated
  directly.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-generate}{\code{ObservationGenerator$generate()}}
\item \href{#method-clone}{\code{ObservationGenerator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate"></a>}}
\if{latex}{\out{\hypertarget{method-generate}{}}}
\subsection{Method \code{generate()}}{
A class that inherits from ObservationGenerator must implement a 
  "generate" method that will create the synthetic observations. 

  This method declaration will cause a program to fail if 
  the method is called and the implementing class does not override it.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationGenerator$generate(base)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base}}{Basis for the artifical observations}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object representing the synthetic observation
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
