% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObservationGenerator.R
\name{ObservationGeneratorNormalAR1Err}
\alias{ObservationGeneratorNormalAR1Err}
\title{Observation error generator for autocorrelated error}
\description{
A observation generator for residual error with 
  random (normally distributed) and autocorrelated
  components.
}
\section{Super classes}{
\code{\link[inferno:ObservationGenerator]{inferno::ObservationGenerator}} -> \code{\link[inferno:ObservationGeneratorNormalErr]{inferno::ObservationGeneratorNormalErr}} -> \code{ObservationGeneratorNormalAR1Err}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{arCoeff}}{The autoregression coefficient}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ObservationGeneratorNormalAR1Err-new}{\code{ObservationGeneratorNormalAR1Err$new()}}
\item \href{#method-ObservationGeneratorNormalAR1Err-generate}{\code{ObservationGeneratorNormalAR1Err$generate()}}
\item \href{#method-ObservationGeneratorNormalAR1Err-clone}{\code{ObservationGeneratorNormalAR1Err$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObservationGeneratorNormalAR1Err-new"></a>}}
\if{latex}{\out{\hypertarget{method-ObservationGeneratorNormalAR1Err-new}{}}}
\subsection{Method \code{new()}}{
Construct a new instance of the clas
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationGeneratorNormalAR1Err$new(..., arCoeff)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Arguments passed to the constructor of the super class}

\item{\code{arCoeff}}{The autoregression coefficient value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObservationGeneratorNormalAR1Err-generate"></a>}}
\if{latex}{\out{\hypertarget{method-ObservationGeneratorNormalAR1Err-generate}{}}}
\subsection{Method \code{generate()}}{
Generates synthetic observations based on components of
  random (normally distributed) and autocorrelated error.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationGeneratorNormalAR1Err$generate(base)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base}}{Basis for the synthetic observations. Synthetic error
is added to the basis to create the observations.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object representing the synthetic observations
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ObservationGeneratorNormalAR1Err-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ObservationGeneratorNormalAR1Err-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ObservationGeneratorNormalAR1Err$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
