% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_tplotfunctions.R
\name{plot.probability}
\alias{plot.probability}
\title{Plot an object of class "probability"}
\usage{
\method{plot}{probability}(
  p,
  variability = NULL,
  PvsY = NULL,
  legend = "top",
  lty = c(1, 2, 4, 3, 6, 5),
  lwd = 2,
  col = palette(),
  type = NULL,
  alpha.f = 1,
  var.alpha.f = NULL,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  ylim = c(0, NA),
  grid = TRUE,
  add = FALSE,
  ...
)
}
\arguments{
\item{p}{object of class "probability", obtained with \code{\link{Pr}} or \code{\link{tailPr}}.}

\item{variability}{one of the values \code{"quantiles"}, \code{"samples"}, \code{"none"} (equivalent to \code{NA} or \code{FALSE}), or \code{NULL} (default), in which case the variability available in \code{p} is used. This argument chooses how to represent the variability of the probability; see \code{\link{Pr}}. If the requested variability is not available in the object \code{p}, then a warning is issued and no variability is plotted.}

\item{PvsY}{logical or \code{NULL}: should probabilities be plotted against their \code{Y} argument? If \code{NULL}, the argument between \code{Y} and \code{X} having larger number of values is chosen. As many probability curves will be plotted as the number of values of the other argument.}

\item{legend}{one of the values \code{"bottomright"}, \code{"bottom"}, \code{"bottomleft"}, \code{"left"}, \code{"topleft"}, \code{"top"}, \code{"topright"}, \code{"right"}, \code{"center"} (see \code{\link[graphics]{legend}}): plot a legend at that position. A value \code{FALSE} or any other does not plot any legend. Default \code{"top"}.}

\item{alpha.f}{Numeric, default 0.25: opacity of the colours, \code{0} being completely invisible and \code{1} completely opaque.}

\item{var.alpha.f}{Numeric: opacity of the quantile bands or of the samples, \code{0} being completely invisible and \code{1} completely opaque.}

\item{...}{other parameters to be passed to \code{\link{flexiplot}}.}
}
\description{
This \code{\link[base]{plot}} method is a utility to plot probabilities obtained with \code{\link{Pr}} or \code{\link{tailPr}}, as well as their variabilities. The probabilities are plotted either against \code{Y}, with one curve for each value of \code{X}, or vice versa.
}
