% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFsamples.R
\name{plotFsamples}
\alias{plotFsamples}
\title{Plot one-dimensional posterior probabilities}
\usage{
plotFsamples(
  file,
  learnt,
  data,
  plotprobability = TRUE,
  plotvariability = "samples",
  nFsamples = NULL,
  datahistogram = !(missing(data) || is.null(data)),
  datascatter = !(missing(data) || is.null(data)),
  parallel = TRUE,
  silent = FALSE
)
}
\arguments{
\item{file}{string: name of plot output file}

\item{data}{data.table object or filepath: datapoints}

\item{plotprobability}{logical: plot the resulting probability curve}

\item{plotvariability}{string, either 'samples' or 'quantiles':
how to plot the variability of the probability distribution with new samples}

\item{nFsamples}{positive number: if plotvariability='samples', then
number of samples of representative frequency distributions to display
as variability;
if plotvariability='quantiles', then
the quantiles (in range 0 to 0.5) to show}

\item{datahistogram}{logical: plot the data as histogram?}

\item{datascatter}{logical: plot the data as scatterplot along the x-axis?}

\item{parallel}{Bool or numeric: whether to use pre-existing parallel
workers, or how many to create and use}

\item{silent}{logical: give warnings or updates in the computation}

\item{learned}{Either a string with the name of a directory or full
path for an 'learnt.rds' object, or such an object itself}
}
\value{
A list with the mutual information, its error, and its unit
}
\description{
Plot one-dimensional posterior probabilities
}
