% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_tplotfunctions.R
\name{plotquantiles}
\alias{plotquantiles}
\title{Plot pairs of quantiles}
\usage{
plotquantiles(x, y, alpha.f = 0.25, border = NA, col = palette()[1], ...)
}
\arguments{
\item{x}{Numeric or character: Set of values that were used as \code{Y} argument to \code{\link{Pr}} (typically they make sense if \code{Y} only was a single variate rather than aconjunction of several).}

\item{y}{Numeric: a matrix having as many rows as \code{x} and an even number of columns, with one column per quantile. Typically these quantiles have been obtained with \code{\link{Pr}}, as its \verb{$quantiles} value. This value is a three-dimensional array, and one of its columns (2nd dimension, corresponding to the possible values of the \code{X} argument of \code{\link{Pr}}) should be selected before being used as \code{y} input.}
}
\description{
Utility function to plot pair of quantiles obtained with \code{\link{Pr}}.
}
