% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pr.R
\name{Pr}
\alias{Pr}
\title{Calculate posterior probabilities}
\usage{
Pr(
  Y,
  X = NULL,
  learnt,
  tails = NULL,
  priorY = NULL,
  nsamples = "all",
  quantiles = c(0.055, 0.25, 0.75, 0.945),
  parallel = NULL,
  silent = FALSE,
  usememory = TRUE,
  keepYX = TRUE
)
}
\arguments{
\item{Y}{Matrix or data.table: set of values of variates of which we want
the joint probability of. One variate per column, one set of values per row.}

\item{X}{Matrix or data.table or \code{NULL}: set of values of variates on which we want to condition the joint probability of \code{Y}. If \code{NULL} (default), no conditioning is made (except for conditioning on the learning dataset and prior assumptions). One variate per column, one set of values per row.}

\item{learnt}{Either a character with the name of a directory or full path for a 'learnt.rds' object, produced by the \code{\link[=learn]{learn()}} function, or such an object itself.}

\item{tails}{Named vector or list, or \code{NULL} (default). The names must match some or all of the variates in arguments \code{Y} and \code{X}. For variates in this list, the probability arguments are understood in an semi-open interval sense: \verb{Y ≤ y} or \verb{Y ≥ y}, an so on. This is true for variates on the left and on the right of the conditional sign \verb{\\|}. A left-open interval \verb{Y ≤ y} is indicated by the values \code{'<='} or \code{'left'} or \code{-1}; a right-open interval \verb{Y ≥ y} is indicated by the values \code{'>='} or \code{'right'} or \code{+1}. Values \code{NULL}, \code{'=='}, \code{0} indicate that a point value \code{Y = y} (not an interval) should be calculated. \strong{NB}: the semi-open intervals \emph{always} include the given value; this is important for ordinal or rounded variates. For instance, if \code{Y} is an integer variate, then to calculate  \code{P(Y < 3)} you should require \code{P(Y <= 2)}; for this reason we also have that \code{P(Y <= 2)} and  \code{P(Y >= 2)} generally add up to \emph{more} than 1.}

\item{priorY}{Numeric vector with the same length as the rows of \code{Y}, or \code{TRUE}, or \code{NULL} (default): prior probabilities or base rates for the \code{Y} values. If \code{TRUE}, the prior probabilities are assumed to be all equal.}

\item{nsamples}{Integer or \code{NULL} or \code{'all'}: desired number of samples of the variability of the probability for \code{Y}. If \code{NULL}, no samples are reported. If \code{'all'} (or \code{Inf}), all samples obtained by the \code{\link[=learn]{learn()}} function are used. Default \code{'all'}.}

\item{quantiles}{Numeric vector, between 0 and 1, or \code{NULL}: desired quantiles of the variability of the probability for \code{Y}. Default \code{c(0.055, 0.25, 0.75, 0.945)}, that is, the 5.5\%, 25\%, 75\%, 94.5\% quantiles (these are typical quantile values in the Bayesian literature: they give 50\% and 89\% credibility intervals, which correspond to 1 shannons and 0.5 shannons of uncertainty). If \code{NULL}, no quantiles are calculated.}

\item{parallel}{Logical or \code{NULL} or positive integer: \code{TRUE}: use roughly half of available cores; \code{FALSE}: use serial computation; \code{NULL}: don't do anything (use pre-registered condition); integer: use this many cores. Default \code{NULL}}

\item{silent}{Logical: give warnings or updates in the computation?
Default \code{FALSE}.}

\item{usememory}{Logical: save partial results to disc, to avoid crashes?
Default \code{TRUE}.}

\item{keepYX}{Logical, default \code{TRUE}: keep a copy of the \code{Y} and \code{X} arguments in the output? This is used for the plot method.}
}
\value{
A list of class \code{probability}, consisting of the elements \code{values},  \code{quantiles} (possibly \code{NULL}), \code{samples} (possibly \code{NULL}), \code{values.MCaccuracy}, \code{quantiles.MCaccuracy} (possibly \code{NULL}), \code{Y}, \code{X}. Element \code{values}: a matrix with the probabilities P(Y|X,data,assumptions), for all combinations of values of \code{Y} (rows) and \code{X} (columns). Element \code{quantiles}: an array with the variability quantiles (3rd dimension of the array) for such probabilities. Element \code{samples}: an array with the variability samples (3rd dimension of the array) for such probabilities. Elements \code{values.MCaccuracy} and \code{quantiles.MCaccuracy}: arrays with the numerical accuracies (roughly speaking a standard deviation) of the Monte Carlo calculations for the \code{values} and \code{quantiles} elements. Elements \code{Y}, \code{X}: copies of the \code{Y} and \code{X} arguments.
}
\description{
This function calculates the posteior probability \code{Pr(Y | X, data)}, where \code{Y} and \code{X} are two (non overlapping) sets of joint variate values. If \code{X} is omitted or \code{NULL}, then the posterior probability \code{Pr(Y | data)} is calculated. The function also gives quantiles about the possible variability of the probability \code{Pr(Y | X, newdata, data)} that we could have if more learning data were provided, as well as a number of samples of the possible values of such probabilities. If several joint values are given for \code{Y} or \code{X}, the function will create a 2D grid of results for all possible compbinations of the given \code{Y} and \code{X} values. This function also allows for base-rate or other prior-probability corrections: If a prior (for instance, a base rate) for \code{Y} is given, the function will calculate the \code{Pr(Y | X, data, prior)} from \code{Pr(X | Y, data)} and the prior by means of Bayes's theorem. Each variate in each argument \code{Y}, \code{X} can be specified either as a point-value \code{Y = y} or as a left-open interval \verb{Y ≤ y} or as a right-open interval \verb{Y ≥ y}, through the argument \code{tails}.
}
