% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pr.R
\name{Pr}
\alias{Pr}
\title{Calculate posterior probabilities}
\usage{
Pr(
  Y,
  X = NULL,
  learnt,
  quantiles = c(0.055, 0.25, 0.75, 0.945),
  nsamples = 100L,
  parallel = TRUE,
  silent = TRUE,
  usememory = TRUE,
  keepYX = TRUE
)
}
\arguments{
\item{Y}{matrix or data.table: set of values of variates of which we want
the joint probability of. One variate per column, one set of values per row.}

\item{X}{matrix or data.table or \code{NULL}: set of values of variates on which we want to condition the joint probability of \code{Y}. If \code{NULL} (default), no conditioning is made (except for conditioning on the learning dataset and prior assumptions). One variate per column, one set of values per row.}

\item{learnt}{Either a string with the name of a directory or full path for a 'learnt.rds' object, produced by the \code{\link{learn}} function, or such an object itself.}

\item{quantiles}{numeric vector, between 0 and 1, or \code{NULL}: desired quantiles of the variability of the probability for \code{Y}. Default \code{c(0.055, 0.25, 0.75, 0.945)}, that is, the 5.5\%, 25\%, 75\%, 94.5\% quantiles (these are typical quantile values in the Bayesian literature: they give 50\% and 89\% credibility intervals, which correspond to 1 shannons and 0.5 shannons of uncertainty). If \code{NULL}, no quantiles are calculated.}

\item{nsamples}{integer or \code{NULL} or \code{"all"}: desired number of samples of the variability of the probability for \code{Y}. If \code{NULL}, no samples are reported. If \code{"all"} (or \code{Inf}), all samples obtained by the \code{\link{learn}} function are used. Default \code{100}.}

\item{parallel}{logical or integer: whether to use pre-existing parallel
workers, or how many to create and use. Default \code{TRUE}.}

\item{silent}{logical: give warnings or updates in the computation?
Default \code{FALSE}.}

\item{usememory}{logical: save partial results to disc, to avoid crashes?
Default \code{TRUE}.}

\item{keepYX}{logical, default \code{TRUE}: keep a copy of the \code{Y} and \code{X} arguments in the output? This is used for the plot method.}
}
\value{
A list of class \code{probability}, consisting of the elements \code{values}, and possibly the elements \code{quantiles} (if non-\code{NULL} argument \code{quantiles}), \code{samples} (if non-\code{NULL} argument \code{nsample}), \code{Y}, \code{X}. Element \code{values}: a matrix with the probabilities P(Y|X,data,assumptions), for all combinations of values of \code{Y} (rows) and \code{X} (columns). Element \code{quantiles}: an array with the variability quantiles (3rd dimension of the array) for such probabilities. Element \code{samples}: an array with the variability samples (3rd dimension of the array) for such probabilities. Elements \code{Y}, \code{X}: copies of the \code{Y} and \code{X} arguments.
}
\description{
This function calculates the probability P(Y | X, data), where Y and X are two (non overlapping) sets of joint variates. The function also gives quantiles about the possible variability of the probability P(Y | X, newdata, data) that we could have if more learning data were provided, as well as a number of samples of the possible values of such probabilities. If several joint values are given for Y or X, the function will create a 2D grid of results for all possible compbinations of the given Y and X values.
}
