% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qPr.R
\name{qPr}
\alias{qPr}
\title{Calculate quantiles}
\usage{
qPr(
  p = c(0.055, 0.5, 0.945),
  Yname,
  X = NULL,
  learnt,
  tails = NULL,
  priorY = NULL,
  nsamples = "all",
  quantiles = c(0.055, 0.5, 0.945),
  parallel = NULL,
  silent = FALSE,
  keepYX = TRUE,
  tol = .Machine$double.eps * 10
)
}
\arguments{
\item{p}{Numeric vector of probability levels. Default: \code{c(0.055, 0.5, 0.945)}.}

\item{Yname}{Character vector: name of variate whose quantiles will be computed.}

\item{X}{Matrix or data.table or \code{NULL} (default): set of values of variates on which we want to condition. If \code{NULL}, no conditioning is made (except for conditioning on the learning dataset and prior assumptions). One variate per column, one set of values per row.}

\item{learnt}{Either a character with the name of a directory or full path for a 'learnt.rds' object, produced by the \code{\link[=learn]{learn()}} function, or such an object itself.}

\item{tails}{Named vector or list, or \code{NULL} (default). The names must match some or all of the variates in arguments \code{X}. For variates in this list, the probability conditional is understood in an semi-open interval sense: \verb{X ≤ x} or \verb{X ≥ x}, an so on. See analogous argument in \code{\link[=Pr]{Pr()}}.}

\item{priorY}{Numeric vector with the same length as the rows of \code{Y}, or \code{TRUE}, or \code{NULL} (default): prior probabilities or base rates for the \code{Y} values. If \code{TRUE}, the prior probabilities are assumed to be all equal. For the moment only the value \code{NULL} is accepted.}

\item{nsamples}{Integer or \code{NULL} or \code{'all'} (default): desired number of samples of the variability of the quantile for \code{Y}. If \code{NULL}, no samples are reported. If \code{'all'} (or \code{Inf}), all samples obtained by the \code{\link[=learn]{learn()}} function are used.}

\item{quantiles}{Numeric vector, between 0 and 1, or \code{NULL}: desired quantiles of the variability of the quantile for \code{Y}. Default \code{c(0.055, 0.25, 0.75, 0.945)}, that is, the 5.5\%, 25\%, 75\%, 94.5\% quantiles (these are typical quantile values in the Bayesian literature: they give 50\% and 89\% credibility intervals, which correspond to 1 shannons and 0.5 shannons of uncertainty). If \code{NULL}, no quantiles are calculated.}

\item{parallel}{Logical or positive integer or cluster object. \code{TRUE}: use roughly half of available cores; \code{FALSE}: use serial computation; integer: use this many cores. It can also be a cluster object previously created with \code{\link[parallel:makeCluster]{parallel::makeCluster()}}; in this case the parallel computation will use this object.}

\item{silent}{Logical, default \code{FALSE}: give warnings or updates in the computation?}

\item{keepYX}{Logical, default \code{TRUE}: keep a copy of the \code{Y} and \code{X} arguments in the output? This is used for the plot method.}

\item{tol}{numeric positive: tolerance in the calculation of quantiles. Default: \code{.Machine$double.eps * 10} (typically \code{2.22045e-15}).}
}
\value{
A list of the elements \code{values},  \code{quantiles} (possibly \code{NULL}), \code{samples} (possibly \code{NULL}), \code{Y}, \code{X}. Element \code{values}: a matrix with the requested \code{Y}-quantiles conditional on the requested \code{X}-values, for all combinations of \code{p} (rows) and \code{X} (columns). Element \code{quantiles}: an array with the variability quantiles (3rd dimension of the array). Element \code{samples}: an array with the variability samples (3rd dimension of the array). Elements \code{Y}, \code{X}: copies of the \code{p} and \code{X} arguments.
}
\description{
This function calculates the quantiles of \code{Pr(Y | X, data)} at specified probability levels, as well as the variability of those quantiles if more learning data were provided. The variability can be expressed in the form of quantiles, samples, or both, as in the \code{\link[=Pr]{Pr()}} function. If several joint values are given for the probability levels and for \code{X}, the function creates a 2D grid of results for all possible compbinations of the given probability levels and \code{X} values. Each variate in the argument \code{X} can be specified either as a point-value \code{X = x} or as a left-open interval \verb{X ≤ x} or as a right-open interval \verb{X ≥ x}, through the argument \code{tails}.
}
