% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayes.R
\name{StatsLoggerBayes}
\alias{StatsLoggerBayes}
\title{R6 class defining a statistics logger for a Bayesian objective function}
\description{
Extends \code{AdaptiveMCMCStatsLogger} to include stats specific to a Bayesian likelihood.
  See \code{\link{AdaptiveMCMCStatsLogger}} for documentation on inherited constructor.
}
\section{Super class}{
\code{\link[inferno:AdaptiveMCMCStatsLogger]{inferno::AdaptiveMCMCStatsLogger}} -> \code{StatsLoggerBayes}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{likelihoodColumn}}{Column index for the likelihood statistic}

\item{\code{propLikelihoodColumn}}{Column index for the proposed likelihood statistic}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StatsLoggerBayes-buildLog}{\code{StatsLoggerBayes$buildLog()}}
\item \href{#method-StatsLoggerBayes-logAccepted}{\code{StatsLoggerBayes$logAccepted()}}
\item \href{#method-StatsLoggerBayes-logRejected}{\code{StatsLoggerBayes$logRejected()}}
\item \href{#method-StatsLoggerBayes-clone}{\code{StatsLoggerBayes$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="inferno" data-topic="AdaptiveMCMCStatsLogger" data-id="initialize"><a href='../../inferno/html/AdaptiveMCMCStatsLogger.html#method-AdaptiveMCMCStatsLogger-initialize'><code>inferno::AdaptiveMCMCStatsLogger$initialize()</code></a></li>
<li><span class="pkg-link" data-pkg="inferno" data-topic="AdaptiveMCMCStatsLogger" data-id="logFailedPredictions"><a href='../../inferno/html/AdaptiveMCMCStatsLogger.html#method-AdaptiveMCMCStatsLogger-logFailedPredictions'><code>inferno::AdaptiveMCMCStatsLogger$logFailedPredictions()</code></a></li>
<li><span class="pkg-link" data-pkg="inferno" data-topic="AdaptiveMCMCStatsLogger" data-id="writeFirstRow"><a href='../../inferno/html/AdaptiveMCMCStatsLogger.html#method-AdaptiveMCMCStatsLogger-writeFirstRow'><code>inferno::AdaptiveMCMCStatsLogger$writeFirstRow()</code></a></li>
<li><span class="pkg-link" data-pkg="inferno" data-topic="AdaptiveMCMCStatsLogger" data-id="writeRow"><a href='../../inferno/html/AdaptiveMCMCStatsLogger.html#method-AdaptiveMCMCStatsLogger-writeRow'><code>inferno::AdaptiveMCMCStatsLogger$writeRow()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatsLoggerBayes-buildLog"></a>}}
\if{latex}{\out{\hypertarget{method-StatsLoggerBayes-buildLog}{}}}
\subsection{Method \code{buildLog()}}{
Overrides the superclass buildLog method to add the structures needed
  for tracking Bayesian specific statistics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatsLoggerBayes$buildLog(numRows, objFunc, filePath)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{numRows}}{Number of rows of data expected in the log}

\item{\code{objFunc}}{The objective function with the statistics to be logged}

\item{\code{filePath}}{The file path where log files should be created}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatsLoggerBayes-logAccepted"></a>}}
\if{latex}{\out{\hypertarget{method-StatsLoggerBayes-logAccepted}{}}}
\subsection{Method \code{logAccepted()}}{
Create accepted log entries
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatsLoggerBayes$logAccepted(index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{The row index for the stats to be logged}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatsLoggerBayes-logRejected"></a>}}
\if{latex}{\out{\hypertarget{method-StatsLoggerBayes-logRejected}{}}}
\subsection{Method \code{logRejected()}}{
Create rejected log entries
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatsLoggerBayes$logRejected(index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{The row index for the stats to be logged}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No defined return value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StatsLoggerBayes-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StatsLoggerBayes-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StatsLoggerBayes$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
