% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_util_lprob.R
\name{util_lprobs}
\alias{util_lprobs}
\title{Calculate collection of log-probabilities for different components and samples}
\usage{
util_lprobs(
  nV0,
  V0mean,
  V0sd,
  xV0,
  nV1,
  V1mean,
  V1sd,
  xV1,
  nV2,
  V2mean,
  V2sd,
  V2steps,
  xV2,
  nVN,
  VNprobs,
  xVN,
  nVB,
  VBprobs,
  xVB
)

util_lprobs(
  nV0,
  V0mean,
  V0sd,
  xV0,
  nV1,
  V1mean,
  V1sd,
  xV1,
  nV2,
  V2mean,
  V2sd,
  V2steps,
  xV2,
  nVN,
  VNprobs,
  xVN,
  nVB,
  VBprobs,
  xVB
)
}
\arguments{
\item{X}{Numerical matrix: transformed variates}

\item{learnt:}{Monte-Carlo-output object}

\item{nR}{etc: Parameters containing appropriate indices}
}
\value{
Matrix with as many rows as components and as many cols as samples

Matrix with as many rows as components and as many cols as samples
}
\description{
Calculate collection of log-probabilities for different components and samples

Calculate collection of log-probabilities for different components and samples
}
\keyword{internal}
