% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_Pcheckpoints.R
\name{util_Pcheckpoints}
\alias{util_Pcheckpoints}
\title{Calculate joint frequencies for checkpoints in learn()}
\usage{
util_Pcheckpoints(Y, mcsamples, auxmetadata)
}
\arguments{
\item{Y}{matrix or data.table: values of some already-transformed
variates of which we want the joint probability; one variate per column}

\item{mcsamples}{object internal to \code{learn()},
containing partial Monte Carlo draws}

\item{auxmetadata}{object internal to \code{learn()},
containing processed metadata information}
}
\value{
The joint frequencies of Y correspoinding to the Monte Carlo samples
}
\description{
Calculate joint frequencies for checkpoints in learn()
}
\keyword{internal}
