% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_checkpoints.R
\name{util_prepPcheckpoints}
\alias{util_prepPcheckpoints}
\title{Format datapoints for testing of MCMC progress}
\usage{
util_prepPcheckpoints(x, auxmetadata, pointsid = NULL)
}
\arguments{
\item{x}{Datapoints to be used for checking MCMC progress}

\item{auxmetadata}{auxmetadata object}

\item{pointsid}{Id of datapoints}
}
\value{
some arguments to be repeatedly used in util_Pcheckpoints
}
\description{
Format datapoints for testing of MCMC progress
}
\keyword{internal}
