% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrtgrid.R
\name{vrtgrid}
\alias{vrtgrid}
\title{Create a grid of values for a variate}
\usage{
vrtgrid(vrt, learnt, length.out = 129)
}
\arguments{
\item{vrt}{Character: name of the variate, must match one of the names in the \code{metadata} file provided to the \code{\link[=learn]{learn()}} function.}

\item{learnt}{Either a character with the name of a directory or full path for a 'learnt.rds' object, produced by the \code{\link[=learn]{learn()}} function, or such an object itself.}

\item{length.out}{Numeric, positive (default 129): number of values to be created; used only for continuous, non-rounded variates (see \code{\link{metadata}}).}
}
\value{
A numeric or character vector of values.
}
\description{
This function create a set of values for a variate \code{vrt}, based on the metadata stored in a \code{learnt} object created by the \code{\link[=learn]{learn()}} function). The set of values depends on the type of variate (nominal or continuous, rounded, and so on, see \code{\link{metadata}}). The range of values is chosen to include, and extend slightly beyond, the range observed in the data used in the \code{\link[=learn]{learn()}} function. Variate domains are always respected.
}
